refactor: decompose TenantCluster controller into sub-reconciler files
Split 3,572-line monolithic controller into 10 focused files:

- tenantcluster_controller.go (363 lines): Reconcile loop, struct, setup
- helpers.go: config resolution, kubeconfig, tenant client, exposure IP
- reconcile_infrastructure.go: CAPI resource creation, status checks, Kamaji compat, image sync
- reconcile_addons.go: addon installation, observability auto-enroll
- reconcile_scaling.go: replica sync, providerID, SCP drift, machine template
- reconcile_talos.go: bootstrap, config apply, talosconfig
- reconcile_ipam.go: IP allocation, elastic IPAM, MetalLB pool updates
- reconcile_deletion.go: CAPI Cluster deletion ordering, namespace cleanup
- reconcile_validation.go: multi-tenancy mode, provider access, quotas
- reconcile_namespace.go: tenant namespace, team RBAC

Pure move. No logic changes, no renames, no refactoring. Every function
retains its exact implementation. Same package, same receiver type.
