+++ internal/controller/tenantcluster/reconcile_talos.go
@@ -1,0 +1,421 @@
+// Copyright 2026 The Butler Authors.
+// SPDX-License-Identifier: Apache-2.0
+
+package tenantcluster
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/log"
+
+	butlerv1alpha1 "github.com/butlerdotdev/butler-api/api/v1alpha1"
+	"github.com/butlerdotdev/butler-controller/internal/capi"
+	"github.com/butlerdotdev/butler-controller/internal/talos"
+)
+
+// reconcileTalosApplyConfig applies Talos machine configs to CAPI Machines in maintenance mode.
+// It discovers Machine IPs, checks whether config was already applied (via annotation),
+// and runs talosctl apply-config --insecure for each unconfigured Machine.
+// reconcileTalosApplyConfig applies Talos machine config to CAPI Machines via talosctl.
+// Returns (allApplied, error) where allApplied is true when all Machines with IPs
+// have their config applied (or are already configured).
+func (r *Reconciler) reconcileTalosApplyConfig(ctx context.Context, tc *butlerv1alpha1.TenantCluster) (bool, error) {
+	logger := log.FromContext(ctx)
+
+	tenantNS := tc.Status.TenantNamespace
+	if tenantNS == "" {
+		return false, nil
+	}
+
+	// Read bootstrap Secret to get the machine config
+	secretName := fmt.Sprintf("%s-talos-bootstrap", tc.Name)
+	bootstrapSecret := &corev1.Secret{}
+	if err := r.Get(ctx, types.NamespacedName{Name: secretName, Namespace: tenantNS}, bootstrapSecret); err != nil {
+		if apierrors.IsNotFound(err) {
+			return false, nil // Bootstrap Secret not ready yet
+		}
+		return false, fmt.Errorf("failed to get bootstrap Secret: %w", err)
+	}
+
+	machineConfig := bootstrapSecret.Data["value"]
+	if len(machineConfig) == 0 {
+		return false, fmt.Errorf("bootstrap Secret %s has empty machine config", secretName)
+	}
+
+	// Get CAPI Machine addresses
+	machines, err := talos.GetMachineAddresses(ctx, r.Client, tc.Name, tenantNS)
+	if err != nil {
+		return false, fmt.Errorf("getting Machine addresses: %w", err)
+	}
+	if len(machines) == 0 {
+		logger.Info("waiting for CAPI Machines to get IP addresses")
+		return false, nil
+	}
+
+	// Apply config to each Machine that hasn't been configured yet
+	talosClient := talos.NewClient()
+	var failCount, pendingCount int
+	for _, machine := range machines {
+		if machine.Annotations != nil && machine.Annotations["butler.butlerlabs.dev/talos-config-applied"] == "true" {
+			continue
+		}
+
+		pendingCount++
+		hostnamePatch := fmt.Sprintf(`{"machine":{"network":{"hostname":%q}}}`, machine.Name)
+		logger.Info("applying Talos config to Machine", "machine", machine.Name, "ip", machine.IP)
+		if err := talosClient.ApplyConfig(ctx, machine.IP, machineConfig, hostnamePatch); err != nil {
+			// "certificate required" means the node left maintenance mode and already has config.
+			// Treat as success and annotate.
+			if strings.Contains(err.Error(), "certificate required") {
+				logger.Info("node already configured (requires TLS client cert), marking as applied",
+					"machine", machine.Name, "ip", machine.IP)
+			} else {
+				logger.Error(err, "failed to apply Talos config", "machine", machine.Name, "ip", machine.IP)
+				failCount++
+				continue
+			}
+		}
+
+		// Annotate the Machine to mark config as applied
+		if err := r.annotateMachine(ctx, machine.Name, tenantNS); err != nil {
+			logger.Error(err, "failed to annotate Machine after config apply", "machine", machine.Name)
+		}
+		pendingCount--
+	}
+
+	if failCount > 0 && failCount == len(machines) {
+		return false, fmt.Errorf("failed to apply Talos config to all %d Machines", failCount)
+	}
+
+	allApplied := pendingCount == 0 && failCount == 0
+	return allApplied, nil
+}
+
+// annotateMachine adds the talos-config-applied annotation to a CAPI Machine.
+func (r *Reconciler) annotateMachine(ctx context.Context, machineName, namespace string) error {
+	machine := &unstructured.Unstructured{}
+	machine.SetGroupVersionKind(schema.GroupVersionKind{
+		Group:   capi.ClusterAPIGroup,
+		Version: capi.ClusterAPIVersion,
+		Kind:    "Machine",
+	})
+	if err := r.Get(ctx, types.NamespacedName{Name: machineName, Namespace: namespace}, machine); err != nil {
+		return err
+	}
+
+	patch := []byte(`{"metadata":{"annotations":{"butler.butlerlabs.dev/talos-config-applied":"true"}}}`)
+	return r.Patch(ctx, machine, client.RawPatch(types.MergePatchType, patch))
+}
+
+// reconcileTalosconfig creates a talosconfig Secret for CLI access to worker nodes.
+// This requires admin client credentials from the trustd-creds Secret (generated by steward).
+func (r *Reconciler) reconcileTalosconfig(ctx context.Context, tc *butlerv1alpha1.TenantCluster) error {
+	logger := log.FromContext(ctx)
+
+	tenantNS := tc.Status.TenantNamespace
+	talosconfigName := fmt.Sprintf("%s-talosconfig", tc.Name)
+
+	// Check if talosconfig Secret already exists
+	existing := &corev1.Secret{}
+	if err := r.Get(ctx, types.NamespacedName{Name: talosconfigName, Namespace: tenantNS}, existing); err == nil {
+		return nil // Already exists
+	}
+
+	// Read trustd credentials
+	trustdCredsName := fmt.Sprintf("%s-trustd-creds", tc.Name)
+	trustdCreds := &corev1.Secret{}
+	if err := r.Get(ctx, types.NamespacedName{Name: trustdCredsName, Namespace: tenantNS}, trustdCreds); err != nil {
+		return fmt.Errorf("failed to get trustd credentials: %w", err)
+	}
+
+	adminCert := trustdCreds.Data["admin.crt"]
+	adminKey := trustdCreds.Data["admin.key"]
+	caCert := trustdCreds.Data["os-ca.crt"]
+
+	if len(adminCert) == 0 || len(adminKey) == 0 {
+		logger.V(1).Info("admin credentials not available in trustd-creds, skipping talosconfig")
+		return nil // Steward hasn't been updated yet
+	}
+
+	// Determine the endpoint for the talosconfig
+	var endpoint string
+	machines, err := talos.GetMachineAddresses(ctx, r.Client, tc.Name, tenantNS)
+	if err == nil && len(machines) > 0 {
+		endpoint = machines[0].IP
+	}
+	if endpoint == "" {
+		return nil // No machines with IPs yet
+	}
+
+	// Generate talosconfig YAML
+	talosconfigData, err := talos.GenerateTalosconfig(talos.TalosconfigInput{
+		ContextName: tc.Name,
+		Endpoints:   []string{endpoint},
+		CACert:      caCert,
+		AdminCert:   adminCert,
+		AdminKey:    adminKey,
+	})
+	if err != nil {
+		return fmt.Errorf("generating talosconfig: %w", err)
+	}
+
+	// Create the Secret
+	secret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      talosconfigName,
+			Namespace: tenantNS,
+			Labels: map[string]string{
+				butlerv1alpha1.LabelManagedBy: "butler",
+				butlerv1alpha1.LabelTenant:    tc.Name,
+			},
+		},
+		Data: map[string][]byte{
+			"talosconfig": talosconfigData,
+		},
+	}
+
+	// Skip owner reference — TC is in a different namespace than the Secret.
+	// Labels track ownership instead.
+
+	if err := r.Create(ctx, secret); err != nil {
+		if apierrors.IsAlreadyExists(err) {
+			return nil
+		}
+		return fmt.Errorf("creating talosconfig Secret: %w", err)
+	}
+
+	logger.Info("created talosconfig Secret", "name", talosconfigName, "namespace", tenantNS)
+
+	return nil
+}
+
+// isTalosCluster returns true if the TenantCluster uses Talos OS for workers.
+func isTalosCluster(tc *butlerv1alpha1.TenantCluster) bool {
+	return talos.IsTalosCluster(tc)
+}
+
+// reconcileTalosBootstrap generates the Talos machine config and creates the bootstrap
+// Secret that CAPI's dataSecretName mechanism reads. This is called after the TCP is
+// ready and the control plane is accessible.
+//
+// The Secret contains a Talos v1alpha1 worker machine config with:
+//   - machine.token: trustd token for apid auth with steward-trustd
+//   - cluster.token: kubeadm bootstrap token for kubelet TLS bootstrapping
+//   - machine.ca.crt: OS CA cert for trusting steward-trustd
+//   - cluster.ca.crt: K8s cluster CA for trusting the API server
+func (r *Reconciler) reconcileTalosBootstrap(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) error {
+	logger := log.FromContext(ctx)
+
+	secretName := fmt.Sprintf("%s-talos-bootstrap", tc.Name)
+	tenantNS := tc.Status.TenantNamespace
+
+	// Check if bootstrap Secret already exists
+	existing := &corev1.Secret{}
+	err := r.Get(ctx, types.NamespacedName{Name: secretName, Namespace: tenantNS}, existing)
+	if err == nil {
+		logger.V(1).Info("Talos bootstrap Secret already exists", "name", secretName)
+		return nil
+	}
+	if !apierrors.IsNotFound(err) {
+		return fmt.Errorf("failed to check for bootstrap Secret: %w", err)
+	}
+
+	// Read trustd credentials from the management cluster
+	trustdCredsName := fmt.Sprintf("%s-trustd-creds", tc.Name)
+	trustdCreds := &corev1.Secret{}
+	if err := r.Get(ctx, types.NamespacedName{Name: trustdCredsName, Namespace: tenantNS}, trustdCreds); err != nil {
+		if apierrors.IsNotFound(err) {
+			logger.Info("trustd credentials not ready yet, waiting", "name", trustdCredsName)
+			return nil
+		}
+		return fmt.Errorf("failed to get trustd credentials: %w", err)
+	}
+
+	machineToken := string(trustdCreds.Data["token"])
+	osCACert := string(trustdCreds.Data["os-ca.crt"])
+	if machineToken == "" || osCACert == "" {
+		return fmt.Errorf("trustd credentials missing token or os-ca.crt")
+	}
+
+	// Read cluster CA from the management cluster
+	caSecretName := fmt.Sprintf("%s-ca", tc.Name)
+	caSecret := &corev1.Secret{}
+	if err := r.Get(ctx, types.NamespacedName{Name: caSecretName, Namespace: tenantNS}, caSecret); err != nil {
+		if apierrors.IsNotFound(err) {
+			logger.Info("cluster CA not ready yet, waiting", "name", caSecretName)
+			return nil
+		}
+		return fmt.Errorf("failed to get cluster CA: %w", err)
+	}
+
+	clusterCACert := string(caSecret.Data["ca.crt"])
+	if clusterCACert == "" {
+		return fmt.Errorf("cluster CA Secret missing ca.crt")
+	}
+
+	// Determine control plane endpoint based on service type.
+	// When using LoadBalancer, use the Service's external IP directly so that
+	// both the API server (port 6443) and trustd (port 50001) are reachable
+	// on the same address. Gateway mode TLS passthrough on multiple ports
+	// for the same hostname has known issues (cilium/cilium#42898).
+	var controlPlaneEndpoint string
+
+	effectiveServiceType := r.effectiveServiceType(tc, butlerConfig)
+	if effectiveServiceType == "LoadBalancer" {
+		svc := &corev1.Service{}
+		if err := r.Get(ctx, types.NamespacedName{Name: tc.Name, Namespace: tenantNS}, svc); err != nil {
+			if apierrors.IsNotFound(err) {
+				logger.Info("TCP Service not found yet, waiting")
+				return nil
+			}
+			return fmt.Errorf("failed to get TCP Service: %w", err)
+		}
+		var lbIP string
+		for _, ingress := range svc.Status.LoadBalancer.Ingress {
+			if ingress.IP != "" {
+				lbIP = ingress.IP
+				break
+			}
+			if ingress.Hostname != "" {
+				lbIP = ingress.Hostname
+				break
+			}
+		}
+		if lbIP == "" {
+			logger.Info("TCP Service LoadBalancer IP not assigned yet, waiting")
+			return nil
+		}
+		controlPlaneEndpoint = fmt.Sprintf("https://%s:6443", lbIP)
+		logger.Info("using LoadBalancer IP for Talos control plane endpoint", "ip", lbIP)
+	} else {
+		// Gateway/Ingress mode — use the TCP status endpoint (hostname-based)
+		tcp := &unstructured.Unstructured{}
+		tcp.SetGroupVersionKind(schema.GroupVersionKind{
+			Group:   "steward.butlerlabs.dev",
+			Version: "v1alpha1",
+			Kind:    "TenantControlPlane",
+		})
+		if err := r.Get(ctx, types.NamespacedName{Name: tc.Name, Namespace: tenantNS}, tcp); err != nil {
+			return fmt.Errorf("failed to get TenantControlPlane: %w", err)
+		}
+		tcpEndpoint, found, _ := unstructured.NestedString(tcp.Object, "status", "controlPlaneEndpoint")
+		if !found || tcpEndpoint == "" {
+			logger.Info("TCP controlPlaneEndpoint not set yet, waiting")
+			return nil
+		}
+		controlPlaneEndpoint = talos.EndpointFromTCPStatus(tcpEndpoint)
+		logger.Info("using Gateway endpoint for Talos control plane endpoint", "endpoint", controlPlaneEndpoint)
+	}
+
+	// Get tenant client to create bootstrap token in tenant API server
+	tenantClient, err := r.getTenantClient(ctx, tc)
+	if err != nil {
+		logger.Info("tenant client not available yet, waiting", "error", err)
+		return nil
+	}
+
+	// Reuse existing bootstrap token if one was created by a previous attempt,
+	// otherwise generate a new one. This ensures idempotency when the reconciler
+	// succeeds in creating the token but fails before creating the bootstrap Secret.
+	bootstrapToken, err := talos.FindExistingBootstrapToken(ctx, tenantClient.Clientset)
+	if err != nil {
+		return fmt.Errorf("failed to check for existing bootstrap token: %w", err)
+	}
+
+	if bootstrapToken != "" {
+		logger.Info("reusing existing kubeadm bootstrap token in tenant API server")
+	} else {
+		bootstrapToken, err = talos.GenerateBootstrapToken()
+		if err != nil {
+			return fmt.Errorf("failed to generate bootstrap token: %w", err)
+		}
+
+		if err := talos.CreateBootstrapTokenSecret(ctx, tenantClient.Clientset, bootstrapToken); err != nil {
+			return fmt.Errorf("failed to create bootstrap token in tenant: %w", err)
+		}
+
+		logger.Info("created kubeadm bootstrap token in tenant API server")
+	}
+
+	// Get networking CIDRs
+	podCIDR := "10.244.0.0/16"
+	serviceCIDR := "10.96.0.0/12"
+	if tc.Spec.Networking.PodCIDR != "" {
+		podCIDR = tc.Spec.Networking.PodCIDR
+	}
+	if tc.Spec.Networking.ServiceCIDR != "" {
+		serviceCIDR = tc.Spec.Networking.ServiceCIDR
+	}
+
+	// Get Talos-specific config
+	installDisk := "/dev/vda"
+	var installerImage string
+	if tc.Spec.Workers.MachineTemplate.OS.Talos != nil {
+		if tc.Spec.Workers.MachineTemplate.OS.Talos.InstallDisk != "" {
+			installDisk = tc.Spec.Workers.MachineTemplate.OS.Talos.InstallDisk
+		}
+		installerImage = tc.Spec.Workers.MachineTemplate.OS.Talos.InstallerImage
+	}
+
+	// Generate Talos machine config
+	input := talos.MachineConfigInput{
+		ClusterName:          tc.Name,
+		ControlPlaneEndpoint: controlPlaneEndpoint,
+		ClusterCACert:        clusterCACert,
+		MachineToken:         machineToken,
+		BootstrapToken:       bootstrapToken,
+		OSCACert:             osCACert,
+		PodCIDR:              podCIDR,
+		ServiceCIDR:          serviceCIDR,
+		InstallDisk:          installDisk,
+		InstallerImage:       installerImage,
+	}
+
+	machineConfigData, err := talos.GenerateWorkerConfig(input)
+	if err != nil {
+		return fmt.Errorf("failed to generate Talos machine config: %w", err)
+	}
+
+	// Create the bootstrap Secret in the tenant namespace on the management cluster.
+	// CAPI reads this via dataSecretName on the MachineDeployment.
+	bootstrapSecret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      secretName,
+			Namespace: tenantNS,
+			Labels: map[string]string{
+				butlerv1alpha1.LabelManagedBy: "butler",
+				butlerv1alpha1.LabelTenant:    tc.Name,
+			},
+		},
+		Data: map[string][]byte{
+			"value": machineConfigData,
+		},
+	}
+
+	// Skip owner reference — TC is in butler-tenants namespace but bootstrap Secret
+	// is in the tenant namespace. Cross-namespace owner refs are disallowed.
+	// Labels track ownership instead.
+
+	if err := r.Create(ctx, bootstrapSecret); err != nil {
+		if apierrors.IsAlreadyExists(err) {
+			return nil
+		}
+		return fmt.Errorf("failed to create Talos bootstrap Secret: %w", err)
+	}
+
+	logger.Info("created Talos bootstrap Secret",
+		"name", secretName,
+		"namespace", tenantNS,
+		"endpoint", controlPlaneEndpoint)
+
+	return nil
+}
