+++ internal/controller/tenantcluster/tenantcluster_controller.go
@@ -18,10 +18,7 @@ package tenantcluster
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"gopkg.in/yaml.v3"
-
 	"strings"
 	"time"
 
@@ -29,12 +26,8 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -44,8 +37,6 @@ import (
 
 	butlerv1alpha1 "github.com/butlerdotdev/butler-api/api/v1alpha1"
 	"github.com/butlerdotdev/butler-controller/internal/addons"
-	"github.com/butlerdotdev/butler-controller/internal/capi"
-	"github.com/butlerdotdev/butler-controller/internal/talos"
 	"github.com/butlerdotdev/butler-controller/internal/tenant"
 )
 
@@ -300,2026 +291,6 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 	return ctrl.Result{RequeueAfter: r.calculateRequeueInterval(tc)}, nil
 }
 
-func (r *Reconciler) reconcileInfrastructure(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-
-	// Already ready, nothing to do
-	if tc.Status.Phase == butlerv1alpha1.TenantClusterPhaseReady {
-		return ctrl.Result{}, nil
-	}
-
-	logger.Info("reconciling infrastructure", "tenantNamespace", tc.Status.TenantNamespace)
-
-	providerConfig, err := r.getProviderConfig(ctx, tc)
-	if err != nil {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionInfrastructureReady,
-			metav1.ConditionFalse, ReasonProviderConfigNotFound, err.Error())
-		return ctrl.Result{}, err
-	}
-
-	// Validate provider scope access
-	if err := r.validateProviderAccess(ctx, tc, providerConfig); err != nil {
-		return ctrl.Result{}, err
-	}
-
-	// Reconcile IPAM allocation (if provider has network.mode=ipam)
-	ipamReady, err := r.reconcileIPAllocation(ctx, tc, providerConfig)
-	if err != nil {
-		return ctrl.Result{}, err
-	}
-	if !ipamReady {
-		logger.Info("waiting for IPAM allocation")
-		return ctrl.Result{RequeueAfter: 10 * time.Second}, nil
-	}
-
-	// Build CAPI resources with ButlerConfig for platform-level settings
-	// ButlerConfig provides ControlPlaneExposure settings (LoadBalancer/Ingress/Gateway mode)
-	// which enables auto-enabling tcp-proxy for non-LoadBalancer modes
-	builder := capi.NewBuilder(tc, providerConfig, tc.Status.TenantNamespace).
-		WithButlerConfig(butlerConfig)
-
-	// For Ingress/Gateway modes, get the Ingress controller IP for worker /etc/hosts
-	// CRITICAL: If Ingress/Gateway mode is enabled but the Ingress controller doesn't have
-	// an external IP yet, we MUST wait. Otherwise, KubeadmConfigTemplate will be created
-	// without the /etc/hosts entry, and workers won't be able to resolve the API server hostname.
-	if butlerConfig != nil {
-		mode := butlerConfig.GetControlPlaneExposureMode()
-		hostname := butlerConfig.GetControlPlaneExposureHostname()
-		logger.V(1).Info("ButlerConfig exposure settings",
-			"mode", mode,
-			"hostname", hostname)
-		if mode == butlerv1alpha1.ControlPlaneExposureModeIngress ||
-			mode == butlerv1alpha1.ControlPlaneExposureModeGateway {
-			ingressIP, err := r.getExposureIP(ctx, butlerConfig)
-			if err != nil {
-				logger.Error(err, "failed to get exposure IP")
-				r.setCondition(tc, butlerv1alpha1.TenantClusterConditionInfrastructureReady,
-					metav1.ConditionFalse, ReasonInfraProvisioning, "Waiting for exposure endpoint")
-				if err := r.Status().Update(ctx, tc); err != nil {
-					return ctrl.Result{}, err
-				}
-				return ctrl.Result{RequeueAfter: 10 * time.Second}, nil
-			}
-			if ingressIP == "" {
-				logger.Info("Ingress/Gateway mode requires external IP, waiting")
-				r.setCondition(tc, butlerv1alpha1.TenantClusterConditionInfrastructureReady,
-					metav1.ConditionFalse, ReasonInfraProvisioning, "Waiting for exposure endpoint external IP")
-				if err := r.Status().Update(ctx, tc); err != nil {
-					return ctrl.Result{}, err
-				}
-				return ctrl.Result{RequeueAfter: 10 * time.Second}, nil
-			}
-			logger.Info("Ingress/Gateway mode detected, setting ingressIP on builder", "ingressIP", ingressIP)
-			builder.WithIngressIP(ingressIP)
-		}
-	}
-
-	resourceSet, err := builder.Build()
-	if err != nil {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionInfrastructureReady,
-			metav1.ConditionFalse, ReasonCAPIResourceError, err.Error())
-		return ctrl.Result{}, fmt.Errorf("failed to build CAPI resources: %w", err)
-	}
-
-	for _, resource := range resourceSet.AllResources() {
-		if resource == nil {
-			continue
-		}
-
-		existing := &unstructured.Unstructured{}
-		existing.SetGroupVersionKind(resource.GroupVersionKind())
-
-		err := r.Get(ctx, types.NamespacedName{
-			Name:      resource.GetName(),
-			Namespace: resource.GetNamespace(),
-		}, existing)
-
-		if apierrors.IsNotFound(err) {
-			logger.Info("creating CAPI resource",
-				"kind", resource.GetKind(),
-				"name", resource.GetName(),
-				"namespace", resource.GetNamespace())
-
-			if err := r.Create(ctx, resource); err != nil {
-				if apierrors.IsAlreadyExists(err) {
-					logger.V(1).Info("CAPI resource created by concurrent reconcile",
-						"kind", resource.GetKind(),
-						"name", resource.GetName())
-				} else {
-					return ctrl.Result{}, fmt.Errorf("failed to create %s %s: %w",
-						resource.GetKind(), resource.GetName(), err)
-				}
-			}
-		} else if err != nil {
-			return ctrl.Result{}, fmt.Errorf("failed to get %s %s: %w",
-				resource.GetKind(), resource.GetName(), err)
-		} else {
-			logger.V(1).Info("CAPI resource already exists",
-				"kind", resource.GetKind(),
-				"name", resource.GetName())
-		}
-	}
-
-	infraReady, controlPlaneReady, workersReady, err := r.checkInfrastructureStatus(ctx, tc)
-	if err != nil {
-		logger.Error(err, "failed to check infrastructure status")
-	}
-
-	patched, err := r.handleKamajiHarvesterCompatibility(ctx, tc, butlerConfig)
-	if err != nil {
-		logger.Error(err, "failed to handle Kamaji compatibility")
-	}
-	if patched {
-		logger.Info("patched StewardControlPlane status for Harvester compatibility")
-		return ctrl.Result{Requeue: true}, nil
-	}
-
-	if infraReady {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionInfrastructureReady,
-			metav1.ConditionTrue, "InfrastructureReady", "Infrastructure cluster is ready")
-	} else {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionInfrastructureReady,
-			metav1.ConditionFalse, ReasonInfraProvisioning, "Infrastructure cluster is provisioning")
-	}
-
-	if workersReady {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionWorkersReady,
-			metav1.ConditionTrue, ReasonWorkersReady, "Workers are ready")
-	} else {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionWorkersReady,
-			metav1.ConditionFalse, ReasonWorkersProvisioning, "Workers are provisioning")
-	}
-
-	if controlPlaneReady {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionControlPlaneReady,
-			metav1.ConditionTrue, ReasonControlPlaneReady, "Control plane is ready")
-
-		// For Talos clusters, create the bootstrap Secret and apply config to workers
-		// as soon as CP is ready. Bootstrap MUST happen here because CAPI Machines
-		// reference the Secret via dataSecretName and block until it exists.
-		// Config must be applied BEFORE addon installation — workers need to leave
-		// maintenance mode and join the cluster before Cilium and other addons can
-		// schedule pods on them.
-		if isTalosCluster(tc) {
-			if err := r.reconcileTalosBootstrap(ctx, tc, butlerConfig); err != nil {
-				logger.Error(err, "failed to reconcile Talos bootstrap")
-			}
-			allApplied, err := r.reconcileTalosApplyConfig(ctx, tc)
-			if err != nil {
-				logger.Error(err, "failed to apply Talos config to workers")
-			}
-			if !allApplied {
-				logger.Info("waiting for Talos config to be applied to all workers before installing addons")
-				return ctrl.Result{RequeueAfter: 15 * time.Second}, nil
-			}
-		}
-
-		// Install addons after workers have config applied (Talos) or as soon as
-		// control plane is accessible (non-Talos). Skip if already ready.
-		if tc.Status.Phase == butlerv1alpha1.TenantClusterPhaseReady {
-			return ctrl.Result{}, nil
-		}
-
-		// Check if control plane is accessible by trying to get kubeconfig
-		_, err := r.getTenantKubeconfig(ctx, tc, butlerConfig)
-		if err == nil {
-			logger.Info("control plane accessible, proceeding to addon installation")
-			return r.reconcileAddons(ctx, tc, butlerConfig)
-		}
-		logger.V(1).Info("waiting for control plane kubeconfig", "error", err)
-	} else {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionControlPlaneReady,
-			metav1.ConditionFalse, ReasonInfraProvisioning, "Control plane is provisioning")
-	}
-
-	if !infraReady || !controlPlaneReady || !workersReady {
-		return ctrl.Result{RequeueAfter: 30 * time.Second}, nil
-	}
-
-	return ctrl.Result{}, nil
-}
-
-// reconcileAddons installs required addons for a functional tenant cluster.
-// These are infrastructure requirements, not optional features.
-// Addons are installed monotonically - they are added but never removed via spec changes.
-func (r *Reconciler) reconcileAddons(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-
-	if tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseInstalling {
-		tc.Status.Phase = butlerv1alpha1.TenantClusterPhaseInstalling
-		if err := r.Status().Update(ctx, tc); err != nil {
-			return ctrl.Result{}, err
-		}
-	}
-
-	kubeconfigData, err := r.getTenantKubeconfig(ctx, tc, butlerConfig)
-	if err != nil {
-		logger.Error(err, "failed to get tenant kubeconfig")
-		return ctrl.Result{RequeueAfter: 30 * time.Second}, nil
-	}
-
-	var addonStatuses []butlerv1alpha1.AddonStatus
-
-	// 1. CNI - REQUIRED, nodes won't be Ready without it
-	ciliumVersion := addons.DefaultCiliumVersion
-	if tc.Spec.Addons.CNI != nil && tc.Spec.Addons.CNI.Version != "" {
-		ciliumVersion = tc.Spec.Addons.CNI.Version
-	}
-
-	// Extract API server endpoint from kubeconfig (used for default/LoadBalancer mode)
-	apiServerHost, apiServerPort := extractAPIServerEndpoint(kubeconfigData)
-
-	// Determine if we need hostAlias for Ingress/Gateway mode
-	// When using Ingress mode with TLS passthrough, Cilium must connect to the
-	// hostname (not IP) so SNI is sent for routing. But CoreDNS needs CNI to start,
-	// creating a chicken-and-egg problem. We solve this with hostAlias.
-	var ingressIP string
-	if butlerConfig != nil {
-		mode := butlerConfig.GetControlPlaneExposureMode()
-		if mode == butlerv1alpha1.ControlPlaneExposureModeIngress ||
-			mode == butlerv1alpha1.ControlPlaneExposureModeGateway {
-			// Get the exposure IP (Gateway or Ingress controller) to add as hostAlias
-			ip, err := r.getExposureIP(ctx, butlerConfig)
-			if err != nil {
-				logger.Error(err, "failed to get exposure IP")
-			}
-			ingressIP = ip
-
-			// For Ingress/Gateway mode, we need the EXTERNAL hostname (from admin.conf)
-			// not the internal .svc endpoint (from admin.svc). The external hostname
-			// is needed for proper SNI routing in TLS passthrough.
-			externalKubeconfig, err := r.getExternalKubeconfig(ctx, tc)
-			if err == nil && externalKubeconfig != nil {
-				apiServerHost, apiServerPort = extractAPIServerEndpoint(externalKubeconfig)
-			}
-			logger.Info("Ingress mode detected", "ingressIP", ingressIP, "externalHost", apiServerHost)
-		}
-	}
-
-	ciliumCfg := addons.CiliumConfig{
-		Version:       ciliumVersion,
-		APIServerHost: apiServerHost,
-		APIServerPort: apiServerPort,
-		IngressIP:     ingressIP,
-	}
-
-	logger.Info("installing Cilium CNI", "version", ciliumVersion,
-		"apiServerHost", apiServerHost, "ingressIP", ingressIP)
-
-	if err := r.Installer.InstallCilium(ctx, kubeconfigData, ciliumCfg); err != nil {
-		logger.Error(err, "failed to install Cilium")
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionAddonsReady,
-			metav1.ConditionFalse, "CNIInstallFailed", err.Error())
-		return ctrl.Result{RequeueAfter: 30 * time.Second}, nil
-	}
-	addonStatuses = append(addonStatuses, butlerv1alpha1.AddonStatus{
-		Name: "cilium", Version: ciliumVersion, Status: "Healthy", ManagedBy: "butler",
-	})
-
-	// 2. cert-manager - needed by many other components
-	certManagerVersion := addons.DefaultCertManagerVersion
-	if tc.Spec.Addons.CertManager != nil && tc.Spec.Addons.CertManager.Version != "" {
-		certManagerVersion = tc.Spec.Addons.CertManager.Version
-	}
-	logger.Info("installing cert-manager", "version", certManagerVersion)
-	if err := r.Installer.InstallCertManager(ctx, kubeconfigData, certManagerVersion); err != nil {
-		logger.Error(err, "failed to install cert-manager")
-	} else {
-		addonStatuses = append(addonStatuses, butlerv1alpha1.AddonStatus{
-			Name: "cert-manager", Version: certManagerVersion, Status: "Healthy", ManagedBy: "butler",
-		})
-	}
-
-	// 3. Longhorn - storage
-	longhornVersion := addons.DefaultLonghornVersion
-	if tc.Spec.Addons.Storage != nil && tc.Spec.Addons.Storage.Version != "" {
-		longhornVersion = tc.Spec.Addons.Storage.Version
-	}
-	logger.Info("installing Longhorn", "version", longhornVersion)
-	if err := r.Installer.InstallLonghorn(ctx, kubeconfigData, longhornVersion); err != nil {
-		logger.Error(err, "failed to install Longhorn")
-	} else {
-		addonStatuses = append(addonStatuses, butlerv1alpha1.AddonStatus{
-			Name: "longhorn", Version: longhornVersion, Status: "Healthy", ManagedBy: "butler",
-		})
-	}
-
-	// 4. MetalLB - LoadBalancer services
-	metallbVersion := addons.DefaultMetalLBVersion
-	if tc.Spec.Addons.LoadBalancer != nil && tc.Spec.Addons.LoadBalancer.Version != "" {
-		metallbVersion = tc.Spec.Addons.LoadBalancer.Version
-	}
-	var poolStart, poolEnd string
-	// Check IPAM allocation first, fall back to manual loadBalancerPool
-	if tc.Status.LBAllocationRef != nil {
-		lbAlloc := &butlerv1alpha1.IPAllocation{}
-		if err := r.Get(ctx, types.NamespacedName{
-			Name:      tc.Status.LBAllocationRef.Name,
-			Namespace: "butler-system",
-		}, lbAlloc); err == nil && lbAlloc.Status.Phase == butlerv1alpha1.IPAllocationPhaseAllocated {
-			poolStart = lbAlloc.Status.StartAddress
-			poolEnd = lbAlloc.Status.EndAddress
-		}
-	}
-	if poolStart == "" && tc.Spec.Networking.LoadBalancerPool != nil {
-		poolStart = tc.Spec.Networking.LoadBalancerPool.Start
-		poolEnd = tc.Spec.Networking.LoadBalancerPool.End
-	}
-	logger.Info("installing MetalLB", "version", metallbVersion, "poolStart", poolStart, "poolEnd", poolEnd)
-	if err := r.Installer.InstallMetalLB(ctx, kubeconfigData, metallbVersion, poolStart, poolEnd); err != nil {
-		logger.Error(err, "failed to install MetalLB")
-	} else {
-		addonStatuses = append(addonStatuses, butlerv1alpha1.AddonStatus{
-			Name: "metallb", Version: metallbVersion, Status: "Healthy", ManagedBy: "butler",
-		})
-	}
-
-	// 5. Traefik - Ingress
-	traefikVersion := addons.DefaultTraefikVersion
-	if tc.Spec.Addons.Ingress != nil && tc.Spec.Addons.Ingress.Version != "" {
-		traefikVersion = tc.Spec.Addons.Ingress.Version
-	}
-	logger.Info("installing Traefik", "version", traefikVersion)
-	if err := r.Installer.InstallTraefik(ctx, kubeconfigData, traefikVersion); err != nil {
-		logger.Error(err, "failed to install Traefik")
-	} else {
-		addonStatuses = append(addonStatuses, butlerv1alpha1.AddonStatus{
-			Name: "traefik", Version: traefikVersion, Status: "Healthy", ManagedBy: "butler",
-		})
-	}
-
-	// Update observed state
-	if tc.Status.ObservedState == nil {
-		tc.Status.ObservedState = &butlerv1alpha1.ObservedClusterState{}
-	}
-	tc.Status.ObservedState.Addons = addonStatuses
-
-	r.setCondition(tc, butlerv1alpha1.TenantClusterConditionAddonsReady,
-		metav1.ConditionTrue, "AddonsInstalled", "All addons installed successfully")
-
-	// Set kubeconfig secret reference for TenantAddon controller
-	tc.Status.KubeconfigSecretRef = &butlerv1alpha1.LocalObjectReference{
-		Name: fmt.Sprintf("%s-admin-kubeconfig", tc.Name),
-	}
-
-	tc.Status.Phase = butlerv1alpha1.TenantClusterPhaseReady
-	now := metav1.Now()
-	tc.Status.LastTransitionTime = &now
-
-	r.setCondition(tc, butlerv1alpha1.TenantClusterConditionReady,
-		metav1.ConditionTrue, ReasonReady, "Cluster is ready")
-	r.Recorder.Eventf(tc, corev1.EventTypeNormal, "ClusterReady",
-		"Cluster %s is ready with %d workers", tc.Name, tc.Spec.Workers.Replicas)
-
-	logger.Info("TenantCluster is ready", "name", tc.Name)
-
-	return ctrl.Result{}, nil
-}
-
-// reconcileAutoEnrollObservability creates TenantAddon resources for observability
-// agents when autoEnroll is enabled in ButlerConfig and a pipeline is configured.
-// Each agent (vector, prometheus, otel) is independently controlled.
-func (r *Reconciler) reconcileAutoEnrollObservability(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-
-	if butlerConfig == nil || butlerConfig.Spec.Observability == nil {
-		return nil
-	}
-
-	obs := butlerConfig.Spec.Observability
-	if obs.Collection == nil || obs.Collection.AutoEnroll == nil {
-		return nil
-	}
-
-	if obs.Pipeline == nil || obs.Pipeline.ClusterRef == nil {
-		return nil
-	}
-
-	// Don't install agents on the pipeline cluster itself
-	pipeRef := obs.Pipeline.ClusterRef
-	if pipeRef.Name == tc.Name && pipeRef.Namespace == tc.Namespace {
-		return nil
-	}
-
-	enroll := obs.Collection.AutoEnroll
-
-	if enroll.VectorAgent && obs.Pipeline.LogEndpoint != "" {
-		if err := r.ensureAutoEnrolledAddon(ctx, tc, "vector-agent", buildVectorAgentValues(obs.Pipeline)); err != nil {
-			logger.Error(err, "failed to auto-enroll vector-agent")
-		}
-	}
-
-	if enroll.Prometheus && obs.Pipeline.MetricEndpoint != "" {
-		if err := r.ensureAutoEnrolledAddon(ctx, tc, "prometheus-operator", buildPrometheusValues(obs.Pipeline)); err != nil {
-			logger.Error(err, "failed to auto-enroll prometheus-operator")
-		}
-	}
-
-	if enroll.OtelCollector && obs.Pipeline.TraceEndpoint != "" {
-		if err := r.ensureAutoEnrolledAddon(ctx, tc, "otel-collector", buildOtelCollectorValues(obs.Pipeline)); err != nil {
-			logger.Error(err, "failed to auto-enroll otel-collector")
-		}
-	}
-
-	return nil
-}
-
-// ensureAutoEnrolledAddon creates a TenantAddon if it doesn't already exist.
-func (r *Reconciler) ensureAutoEnrolledAddon(ctx context.Context, tc *butlerv1alpha1.TenantCluster, addonDefName string, values *butlerv1alpha1.ExtensionValues) error {
-	logger := log.FromContext(ctx)
-	addonName := fmt.Sprintf("%s-%s", tc.Name, addonDefName)
-
-	existing := &butlerv1alpha1.TenantAddon{}
-	err := r.Get(ctx, client.ObjectKey{Name: addonName, Namespace: tc.Namespace}, existing)
-	if err == nil {
-		return nil // already exists
-	}
-	if !apierrors.IsNotFound(err) {
-		return fmt.Errorf("failed to check TenantAddon %s: %w", addonName, err)
-	}
-
-	addon := &butlerv1alpha1.TenantAddon{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      addonName,
-			Namespace: tc.Namespace,
-			Labels: map[string]string{
-				butlerv1alpha1.LabelManagedBy:              "butler",
-				butlerv1alpha1.LabelTeam:                   tc.Labels[butlerv1alpha1.LabelTeam],
-				"butler.butlerlabs.dev/cluster":            tc.Name,
-				"butler.butlerlabs.dev/addon-definition":   addonDefName,
-				"butler.butlerlabs.dev/auto-enrolled":      "true",
-			},
-		},
-		Spec: butlerv1alpha1.TenantAddonSpec{
-			ClusterRef: butlerv1alpha1.LocalObjectReference{Name: tc.Name},
-			Addon:      addonDefName,
-			Version:    "", // use AddonDefinition default
-			Values:     values,
-		},
-	}
-
-	if err := controllerutil.SetControllerReference(tc, addon, r.Scheme); err != nil {
-		return fmt.Errorf("failed to set owner reference: %w", err)
-	}
-
-	logger.Info("auto-enrolling observability addon",
-		"addon", addonName, "cluster", tc.Name)
-
-	if err := r.Create(ctx, addon); err != nil {
-		if apierrors.IsAlreadyExists(err) {
-			return nil
-		}
-		return fmt.Errorf("failed to create TenantAddon %s: %w", addonName, err)
-	}
-
-	return nil
-}
-
-// buildVectorAgentValues configures the vector-agent sink to forward to the pipeline aggregator.
-func buildVectorAgentValues(pipeline *butlerv1alpha1.ObservabilityPipelineConfig) *butlerv1alpha1.ExtensionValues {
-	if pipeline == nil || pipeline.LogEndpoint == "" {
-		return nil
-	}
-
-	values := map[string]interface{}{
-		"customConfig": map[string]interface{}{
-			"sinks": map[string]interface{}{
-				"aggregator": map[string]interface{}{
-					"uri": pipeline.LogEndpoint,
-				},
-			},
-		},
-	}
-
-	raw, err := json.Marshal(values)
-	if err != nil {
-		return nil
-	}
-	return &butlerv1alpha1.ExtensionValues{Raw: raw}
-}
-
-// buildPrometheusValues configures remote-write to the pipeline metric endpoint.
-func buildPrometheusValues(pipeline *butlerv1alpha1.ObservabilityPipelineConfig) *butlerv1alpha1.ExtensionValues {
-	if pipeline == nil || pipeline.MetricEndpoint == "" {
-		return nil
-	}
-
-	values := map[string]interface{}{
-		"prometheus": map[string]interface{}{
-			"prometheusSpec": map[string]interface{}{
-				"remoteWrite": []map[string]interface{}{
-					{"url": pipeline.MetricEndpoint},
-				},
-			},
-		},
-	}
-
-	raw, err := json.Marshal(values)
-	if err != nil {
-		return nil
-	}
-	return &butlerv1alpha1.ExtensionValues{Raw: raw}
-}
-
-// buildOtelCollectorValues configures the OTLP exporter to the pipeline trace endpoint.
-func buildOtelCollectorValues(pipeline *butlerv1alpha1.ObservabilityPipelineConfig) *butlerv1alpha1.ExtensionValues {
-	if pipeline == nil || pipeline.TraceEndpoint == "" {
-		return nil
-	}
-
-	values := map[string]interface{}{
-		"config": map[string]interface{}{
-			"exporters": map[string]interface{}{
-				"otlp": map[string]interface{}{
-					"endpoint": pipeline.TraceEndpoint,
-					"tls": map[string]interface{}{
-						"insecure": true,
-					},
-				},
-			},
-		},
-	}
-
-	raw, err := json.Marshal(values)
-	if err != nil {
-		return nil
-	}
-	return &butlerv1alpha1.ExtensionValues{Raw: raw}
-}
-
-func (r *Reconciler) getTenantKubeconfig(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) ([]byte, error) {
-	secretName := fmt.Sprintf("%s-admin-kubeconfig", tc.Name)
-
-	secret := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{
-		Name:      secretName,
-		Namespace: tc.Status.TenantNamespace,
-	}, secret); err != nil {
-		return nil, fmt.Errorf("failed to get kubeconfig secret: %w", err)
-	}
-
-	// Determine which kubeconfig key to use based on control plane exposure mode
-	// For Ingress/Gateway modes, use admin.svc (internal service endpoint) since the external
-	// hostname isn't resolvable from within the management cluster
-	kubeconfigKey := "admin.conf"
-	if butlerConfig != nil {
-		mode := butlerConfig.GetControlPlaneExposureMode()
-		if mode == butlerv1alpha1.ControlPlaneExposureModeIngress ||
-			mode == butlerv1alpha1.ControlPlaneExposureModeGateway {
-			kubeconfigKey = "admin.svc"
-		}
-	}
-
-	kubeconfigData, ok := secret.Data[kubeconfigKey]
-	if !ok {
-		// Fallback to admin.conf if preferred key doesn't exist
-		kubeconfigData, ok = secret.Data["admin.conf"]
-		if !ok {
-			return nil, fmt.Errorf("kubeconfig secret missing %s and admin.conf keys", kubeconfigKey)
-		}
-	}
-
-	return kubeconfigData, nil
-}
-
-// getTenantClient returns a cached Kubernetes client for the tenant cluster.
-// Falls back to ad-hoc client creation if the ClientManager is not configured.
-func (r *Reconciler) getTenantClient(ctx context.Context, tc *butlerv1alpha1.TenantCluster) (*tenant.TenantClient, error) {
-	if r.ClientManager == nil {
-		return nil, fmt.Errorf("tenant ClientManager not configured")
-	}
-	if tc.Status.TenantNamespace == "" {
-		return nil, fmt.Errorf("tenant namespace not yet assigned")
-	}
-	return r.ClientManager.GetClient(ctx, tc.Status.TenantNamespace, tc.Name)
-}
-
-// getExternalKubeconfig returns the kubeconfig with the external hostname (admin.conf).
-// This is used for Cilium configuration in Ingress/Gateway mode to ensure proper SNI is sent.
-func (r *Reconciler) getExternalKubeconfig(ctx context.Context, tc *butlerv1alpha1.TenantCluster) ([]byte, error) {
-	secretName := fmt.Sprintf("%s-admin-kubeconfig", tc.Name)
-
-	secret := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{
-		Name:      secretName,
-		Namespace: tc.Status.TenantNamespace,
-	}, secret); err != nil {
-		return nil, fmt.Errorf("failed to get kubeconfig secret: %w", err)
-	}
-
-	// admin.conf contains the external hostname for Ingress/Gateway mode
-	kubeconfigData, ok := secret.Data["admin.conf"]
-	if !ok {
-		return nil, fmt.Errorf("kubeconfig secret missing admin.conf key")
-	}
-
-	return kubeconfigData, nil
-}
-
-func (r *Reconciler) getProviderConfig(ctx context.Context, tc *butlerv1alpha1.TenantCluster) (*butlerv1alpha1.ProviderConfig, error) {
-	const defaultProviderNamespace = "butler-system"
-
-	if tc.Spec.ProviderConfigRef != nil && tc.Spec.ProviderConfigRef.Name != "" {
-		ns := tc.Spec.ProviderConfigRef.Namespace
-		if ns == "" {
-			ns = defaultProviderNamespace
-		}
-		pc := &butlerv1alpha1.ProviderConfig{}
-		if err := r.Get(ctx, types.NamespacedName{Name: tc.Spec.ProviderConfigRef.Name, Namespace: ns}, pc); err != nil {
-			return nil, fmt.Errorf("failed to get ProviderConfig %s/%s: %w", ns, tc.Spec.ProviderConfigRef.Name, err)
-		}
-		return pc, nil
-	}
-
-	butlerConfig := &butlerv1alpha1.ButlerConfig{}
-	if err := r.Get(ctx, types.NamespacedName{Name: ButlerConfigSingletonName}, butlerConfig); err == nil {
-		if butlerConfig.Spec.DefaultProviderConfigRef != nil && butlerConfig.Spec.DefaultProviderConfigRef.Name != "" {
-			pc := &butlerv1alpha1.ProviderConfig{}
-			if err := r.Get(ctx, types.NamespacedName{Name: butlerConfig.Spec.DefaultProviderConfigRef.Name, Namespace: defaultProviderNamespace}, pc); err != nil {
-				return nil, fmt.Errorf("failed to get default ProviderConfig %s/%s from ButlerConfig: %w",
-					defaultProviderNamespace, butlerConfig.Spec.DefaultProviderConfigRef.Name, err)
-			}
-			return pc, nil
-		}
-	}
-
-	pc := &butlerv1alpha1.ProviderConfig{}
-	if err := r.Get(ctx, types.NamespacedName{Name: "default", Namespace: defaultProviderNamespace}, pc); err == nil {
-		return pc, nil
-	}
-
-	pcList := &butlerv1alpha1.ProviderConfigList{}
-	if err := r.List(ctx, pcList, client.InNamespace(defaultProviderNamespace)); err != nil {
-		return nil, fmt.Errorf("failed to list ProviderConfigs: %w", err)
-	}
-
-	if len(pcList.Items) > 0 {
-		return &pcList.Items[0], nil
-	}
-
-	return nil, fmt.Errorf("no ProviderConfig available in %s namespace", defaultProviderNamespace)
-}
-
-// checkInfrastructureStatus checks the status of the CAPI Cluster and MachineDeployment.
-// It updates the TenantCluster status with current worker node counts.
-func (r *Reconciler) checkInfrastructureStatus(ctx context.Context, tc *butlerv1alpha1.TenantCluster) (infraReady, cpReady, workersReady bool, err error) {
-	logger := log.FromContext(ctx)
-
-	cluster := &unstructured.Unstructured{}
-	cluster.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ClusterAPIGroup,
-		Version: capi.ClusterAPIVersion,
-		Kind:    "Cluster",
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Name:      tc.Name,
-		Namespace: tc.Status.TenantNamespace,
-	}, cluster); err != nil {
-		if apierrors.IsNotFound(err) {
-			return false, false, false, nil
-		}
-		return false, false, false, err
-	}
-
-	phase, found, _ := unstructured.NestedString(cluster.Object, "status", "phase")
-	if found {
-		logger.V(1).Info("cluster phase", "phase", phase)
-		infraReady = phase == "Provisioned"
-	}
-
-	cpReadyVal, found, _ := unstructured.NestedBool(cluster.Object, "status", "controlPlaneReady")
-	if found {
-		cpReady = cpReadyVal
-	}
-
-	infraReadyVal, found, _ := unstructured.NestedBool(cluster.Object, "status", "infrastructureReady")
-	if found && !infraReady {
-		infraReady = infraReadyVal
-	}
-
-	md := &unstructured.Unstructured{}
-	md.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ClusterAPIGroup,
-		Version: capi.ClusterAPIVersion,
-		Kind:    "MachineDeployment",
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Name:      fmt.Sprintf("%s-workers", tc.Name),
-		Namespace: tc.Status.TenantNamespace,
-	}, md); err != nil {
-		if apierrors.IsNotFound(err) {
-			return infraReady, cpReady, false, nil
-		}
-		return infraReady, cpReady, false, err
-	}
-
-	replicas, _, _ := unstructured.NestedInt64(md.Object, "status", "replicas")
-	readyReplicas, _, _ := unstructured.NestedInt64(md.Object, "status", "readyReplicas")
-
-	logger.V(1).Info("MachineDeployment status", "replicas", replicas, "readyReplicas", readyReplicas)
-
-	// Update TenantCluster status with worker counts
-	desiredReplicas := int64(tc.Spec.Workers.Replicas)
-	if desiredReplicas < 1 {
-		desiredReplicas = 1
-	}
-	tc.Status.WorkerNodesReady = int32(readyReplicas)
-	tc.Status.WorkerNodesDesired = int32(desiredReplicas)
-
-	// Workers are ready when at least one replica is ready
-	if readyReplicas > 0 {
-		workersReady = true
-	}
-
-	return infraReady, cpReady, workersReady, nil
-}
-
-// reconcileMachineDeploymentReplicas syncs worker count from TenantCluster spec to MachineDeployment.
-// This is called on every reconcile to ensure spec.workers.replicas changes are propagated
-// to the underlying CAPI MachineDeployment, enabling cluster scaling operations.
-// It also updates the TenantCluster status with current worker counts and the WorkersReady condition.
-func (r *Reconciler) reconcileMachineDeploymentReplicas(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-
-	// Only sync when cluster is Ready or Installing
-	// During Provisioning, initial creation handles replicas
-	if tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseReady &&
-		tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseInstalling {
-		return nil
-	}
-
-	// Need tenant namespace to find MachineDeployment
-	if tc.Status.TenantNamespace == "" {
-		return nil
-	}
-
-	// Get the MachineDeployment
-	mdName := fmt.Sprintf("%s-workers", tc.Name)
-	md := &unstructured.Unstructured{}
-	md.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ClusterAPIGroup,
-		Version: capi.ClusterAPIVersion,
-		Kind:    "MachineDeployment",
-	})
-
-	if err := r.Get(ctx, types.NamespacedName{
-		Namespace: tc.Status.TenantNamespace,
-		Name:      mdName,
-	}, md); err != nil {
-		if apierrors.IsNotFound(err) {
-			// Not created yet, will be handled by initial provisioning
-			return nil
-		}
-		return fmt.Errorf("failed to get MachineDeployment: %w", err)
-	}
-
-	// Get current spec replicas and desired replicas from TenantCluster
-	currentSpecReplicas, _, _ := unstructured.NestedInt64(md.Object, "spec", "replicas")
-	desiredReplicas := int64(tc.Spec.Workers.Replicas)
-
-	// Default to 1 if not specified (matches builder.go behavior)
-	if desiredReplicas < 1 {
-		desiredReplicas = 1
-	}
-
-	// Update TenantCluster status with worker counts from MachineDeployment
-	readyReplicas, _, _ := unstructured.NestedInt64(md.Object, "status", "readyReplicas")
-
-	// CAPI MachineDeployment readyReplicas can be 0 even when workers are actually running,
-	// because CAPI can't match Machine→Node without spec.providerID (e.g. Talos, Flatcar on KubeVirt).
-	// When the cluster is already Ready, fall back to querying the tenant API server directly.
-	// To avoid unnecessary API calls at scale, skip the query when the previous reconcile already
-	// found all workers ready (status matches desired). Only re-query on first check or during scaling.
-	if readyReplicas == 0 && tc.Status.Phase == butlerv1alpha1.TenantClusterPhaseReady {
-		previousReady := tc.Status.WorkerNodesReady
-		if previousReady > 0 && previousReady == int32(desiredReplicas) {
-			// Workers fully converged last time — reuse cached count
-			readyReplicas = int64(previousReady)
-		} else {
-			// First check, scaling in progress, or workers not yet converged — query tenant API
-			if nodeCount, err := r.countTenantReadyNodes(ctx, tc, butlerConfig); err == nil && nodeCount > 0 {
-				readyReplicas = int64(nodeCount)
-				logger.V(1).Info("using tenant node count as CAPI readyReplicas fallback",
-					"nodeCount", nodeCount)
-			}
-		}
-	}
-
-	tc.Status.WorkerNodesReady = int32(readyReplicas)
-	tc.Status.WorkerNodesDesired = int32(desiredReplicas)
-
-	// Update WorkersReady condition based on current worker counts
-	if tc.Status.WorkerNodesReady >= tc.Status.WorkerNodesDesired && tc.Status.WorkerNodesDesired > 0 {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionWorkersReady,
-			metav1.ConditionTrue, ReasonWorkersReady,
-			fmt.Sprintf("All %d workers ready", tc.Status.WorkerNodesReady))
-	} else {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionWorkersReady,
-			metav1.ConditionFalse, ReasonWorkersProvisioning,
-			fmt.Sprintf("Workers provisioning: %d/%d ready", tc.Status.WorkerNodesReady, tc.Status.WorkerNodesDesired))
-	}
-
-	// Check if MachineDeployment spec needs to be scaled
-	if currentSpecReplicas == desiredReplicas {
-		// Already in sync, nothing more to do
-		return nil
-	}
-
-	logger.Info("scaling MachineDeployment",
-		"name", mdName,
-		"from", currentSpecReplicas,
-		"to", desiredReplicas)
-
-	// Patch MachineDeployment replicas
-	patch := []byte(fmt.Sprintf(`{"spec":{"replicas":%d}}`, desiredReplicas))
-
-	if err := r.Patch(ctx, md, client.RawPatch(types.MergePatchType, patch)); err != nil {
-		return fmt.Errorf("failed to patch MachineDeployment replicas: %w", err)
-	}
-
-	r.Recorder.Eventf(tc, corev1.EventTypeNormal, "ClusterScaling",
-		"Scaled workers from %d to %d", currentSpecReplicas, desiredReplicas)
-
-	logger.Info("MachineDeployment scaled successfully",
-		"name", mdName,
-		"replicas", desiredReplicas)
-
-	return nil
-}
-
-// countTenantReadyNodes queries the tenant cluster's API server to count Ready worker nodes.
-// This is a fallback for when CAPI readyReplicas is unreliable (e.g. no providerID on nodes).
-// Returns 0 without error if the tenant cluster is unreachable (non-fatal).
-func (r *Reconciler) countTenantReadyNodes(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) (int32, error) {
-	tenantClient, err := r.getTenantClient(ctx, tc)
-	if err != nil {
-		return 0, nil // non-fatal: kubeconfig not available yet
-	}
-
-	nodes, err := tenantClient.Clientset.CoreV1().Nodes().List(ctx, metav1.ListOptions{})
-	if err != nil {
-		return 0, nil // non-fatal: tenant cluster may be unreachable
-	}
-
-	var readyCount int32
-	for _, node := range nodes.Items {
-		// Skip control plane nodes (only count workers)
-		if _, isMaster := node.Labels["node-role.kubernetes.io/master"]; isMaster {
-			continue
-		}
-		if _, isCP := node.Labels["node-role.kubernetes.io/control-plane"]; isCP {
-			continue
-		}
-		for _, cond := range node.Status.Conditions {
-			if cond.Type == corev1.NodeReady && cond.Status == corev1.ConditionTrue {
-				readyCount++
-				break
-			}
-		}
-	}
-
-	return readyCount, nil
-}
-
-// reconcileNodeProviderIDs patches spec.providerID on tenant cluster Nodes to match
-// the CAPI Machine providerID. Without this, CAPI cannot populate Machine.status.nodeRef
-// and MachineDeployment rolling updates stall. Affects Talos and Flatcar workers on
-// KubeVirt and Nutanix where the kubelet does not set providerID natively.
-// No-op for nodes that already have providerID set (including cloud provider nodes).
-func (r *Reconciler) reconcileNodeProviderIDs(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) error {
-	if tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseReady {
-		return nil
-	}
-	if tc.Status.TenantNamespace == "" {
-		return nil
-	}
-
-	logger := log.FromContext(ctx)
-
-	// Short-circuit: if CAPI MachineDeployment already reports all replicas ready,
-	// providerIDs are already set (CAPI needs nodeRef to count ready replicas).
-	mdName := fmt.Sprintf("%s-workers", tc.Name)
-	md := &unstructured.Unstructured{}
-	md.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ClusterAPIGroup,
-		Version: capi.ClusterAPIVersion,
-		Kind:    "MachineDeployment",
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Namespace: tc.Status.TenantNamespace,
-		Name:      mdName,
-	}, md); err != nil {
-		return nil
-	}
-	mdDesired, _, _ := unstructured.NestedInt64(md.Object, "spec", "replicas")
-	mdReady, _, _ := unstructured.NestedInt64(md.Object, "status", "readyReplicas")
-	mdTotal, _, _ := unstructured.NestedInt64(md.Object, "status", "replicas")
-	// Only short-circuit when all replicas are ready AND no extra replicas exist
-	// (rolling updates temporarily create more replicas than desired).
-	if mdDesired > 0 && mdReady == mdDesired && mdTotal == mdDesired {
-		return nil
-	}
-
-	machineList := &unstructured.UnstructuredList{}
-	machineList.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ClusterAPIGroup,
-		Version: capi.ClusterAPIVersion,
-		Kind:    "MachineList",
-	})
-	if err := r.List(ctx, machineList, client.InNamespace(tc.Status.TenantNamespace),
-		client.MatchingLabels{"cluster.x-k8s.io/cluster-name": tc.Name}); err != nil {
-		return fmt.Errorf("failed to list Machines: %w", err)
-	}
-
-	// Build IP -> providerID map. Skip Machines in Deleting/Failed phase and
-	// detect duplicate IPs (can happen during rolling updates) to avoid wrong assignments.
-	ipToProviderID := make(map[string]string)
-	duplicateIPs := make(map[string]bool)
-	for _, machine := range machineList.Items {
-		phase, _, _ := unstructured.NestedString(machine.Object, "status", "phase")
-		if phase == "Deleting" || phase == "Failed" {
-			continue
-		}
-		providerID, _, _ := unstructured.NestedString(machine.Object, "spec", "providerID")
-		if providerID == "" {
-			continue
-		}
-		addresses, _, _ := unstructured.NestedSlice(machine.Object, "status", "addresses")
-		for _, addr := range addresses {
-			addrMap, ok := addr.(map[string]interface{})
-			if !ok {
-				continue
-			}
-			addrType, _ := addrMap["type"].(string)
-			addrValue, _ := addrMap["address"].(string)
-			if addrType != "InternalIP" || addrValue == "" {
-				continue
-			}
-			if _, exists := ipToProviderID[addrValue]; exists {
-				duplicateIPs[addrValue] = true
-				continue
-			}
-			ipToProviderID[addrValue] = providerID
-		}
-	}
-	for ip := range duplicateIPs {
-		logger.V(1).Info("skipping ambiguous IP shared by multiple Machines", "ip", ip)
-		delete(ipToProviderID, ip)
-	}
-
-	if len(ipToProviderID) == 0 {
-		return nil
-	}
-
-	tenantClient, err := r.getTenantClient(ctx, tc)
-	if err != nil {
-		return fmt.Errorf("failed to get tenant client: %w", err)
-	}
-
-	nodes, err := tenantClient.Clientset.CoreV1().Nodes().List(ctx, metav1.ListOptions{})
-	if err != nil {
-		return fmt.Errorf("failed to list tenant nodes: %w", err)
-	}
-
-	for i := range nodes.Items {
-		node := &nodes.Items[i]
-		if node.Spec.ProviderID != "" {
-			continue
-		}
-
-		var nodeIP string
-		for _, addr := range node.Status.Addresses {
-			if addr.Type == corev1.NodeInternalIP {
-				nodeIP = addr.Address
-				break
-			}
-		}
-		if nodeIP == "" {
-			continue
-		}
-
-		providerID, ok := ipToProviderID[nodeIP]
-		if !ok {
-			continue
-		}
-
-		patch := []byte(fmt.Sprintf(`{"spec":{"providerID":%q}}`, providerID))
-		if _, err := tenantClient.Clientset.CoreV1().Nodes().Patch(ctx, node.Name, types.StrategicMergePatchType, patch, metav1.PatchOptions{}); err != nil {
-			logger.Info("failed to set providerID on tenant node", "node", node.Name, "providerID", providerID, "error", err)
-			continue
-		}
-		logger.Info("set providerID on tenant node", "node", node.Name, "providerID", providerID)
-	}
-
-	return nil
-}
-
-// reconcileStewardControlPlane patches the StewardControlPlane when the TenantCluster
-// spec diverges from the live SCP (version, replicas, CP resources).
-func (r *Reconciler) reconcileStewardControlPlane(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-
-	if tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseReady &&
-		tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseInstalling {
-		return nil
-	}
-	if tc.Status.TenantNamespace == "" {
-		return nil
-	}
-
-	scp := &unstructured.Unstructured{}
-	scp.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ControlPlaneAPIGroup,
-		Version: capi.ControlPlaneAPIVersion,
-		Kind:    "StewardControlPlane",
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Namespace: tc.Status.TenantNamespace,
-		Name:      tc.Name,
-	}, scp); err != nil {
-		if apierrors.IsNotFound(err) {
-			return nil
-		}
-		return fmt.Errorf("failed to get StewardControlPlane: %w", err)
-	}
-
-	patch := map[string]interface{}{"spec": map[string]interface{}{}}
-	specPatch := patch["spec"].(map[string]interface{})
-	var changes []string
-
-	// Version drift
-	currentVersion, _, _ := unstructured.NestedString(scp.Object, "spec", "version")
-	if tc.Spec.KubernetesVersion != "" && tc.Spec.KubernetesVersion != currentVersion {
-		specPatch["version"] = tc.Spec.KubernetesVersion
-		changes = append(changes, fmt.Sprintf("version %s->%s", currentVersion, tc.Spec.KubernetesVersion))
-	}
-
-	// Replicas drift. Only patch if TC explicitly sets replicas (> 0).
-	// SCP defaults to 2 via kubebuilder; patching 0 -> 1 would fight the default.
-	currentReplicas, _, _ := unstructured.NestedInt64(scp.Object, "spec", "replicas")
-	desiredReplicas := int64(tc.Spec.ControlPlane.Replicas)
-	if desiredReplicas > 0 && desiredReplicas != currentReplicas {
-		specPatch["replicas"] = desiredReplicas
-		changes = append(changes, fmt.Sprintf("replicas %d->%d", currentReplicas, desiredReplicas))
-	}
-
-	// CP resource drift
-	desired := capi.ResolveControlPlaneResources(tc, butlerConfig)
-	if desired != nil {
-		if desired.APIServer != nil {
-			if cpComponentDrifted(scp, "apiServer", desired.APIServer) {
-				specPatch["apiServer"] = capi.ComponentResourceMap(desired.APIServer)
-				changes = append(changes, "apiServer resources")
-			}
-		}
-		if desired.ControllerManager != nil {
-			if cpComponentDrifted(scp, "controllerManager", desired.ControllerManager) {
-				specPatch["controllerManager"] = capi.ComponentResourceMap(desired.ControllerManager)
-				changes = append(changes, "controllerManager resources")
-			}
-		}
-		if desired.Scheduler != nil {
-			if cpComponentDrifted(scp, "scheduler", desired.Scheduler) {
-				specPatch["scheduler"] = capi.ComponentResourceMap(desired.Scheduler)
-				changes = append(changes, "scheduler resources")
-			}
-		}
-	}
-
-	if len(changes) == 0 {
-		return nil
-	}
-
-	patchBytes, err := json.Marshal(patch)
-	if err != nil {
-		return fmt.Errorf("failed to marshal SCP patch: %w", err)
-	}
-
-	logger.Info("updating StewardControlPlane", "cluster", tc.Name, "changes", changes)
-	return r.Patch(ctx, scp, client.RawPatch(types.MergePatchType, patchBytes))
-}
-
-func cpComponentDrifted(scp *unstructured.Unstructured, component string, desired *butlerv1alpha1.ComponentResources) bool {
-	check := func(path []string, want *resource.Quantity) bool {
-		if want == nil {
-			return false
-		}
-		cur, ok, _ := unstructured.NestedString(scp.Object, path...)
-		if !ok {
-			return true
-		}
-		curQ, err := resource.ParseQuantity(cur)
-		if err != nil {
-			return true
-		}
-		return curQ.Cmp(*want) != 0
-	}
-
-	if desired.Requests != nil {
-		if check([]string{"spec", component, "resources", "requests", "cpu"}, desired.Requests.CPU) {
-			return true
-		}
-		if check([]string{"spec", component, "resources", "requests", "memory"}, desired.Requests.Memory) {
-			return true
-		}
-	}
-	if desired.Limits != nil {
-		if check([]string{"spec", component, "resources", "limits", "cpu"}, desired.Limits.CPU) {
-			return true
-		}
-		if check([]string{"spec", component, "resources", "limits", "memory"}, desired.Limits.Memory) {
-			return true
-		}
-	}
-	return false
-}
-
-// reconcileMachineTemplate detects CPU/memory/disk changes in the TenantCluster spec
-// and performs a rolling update by creating a new immutable MachineTemplate and updating
-// the MachineDeployment's infrastructureRef to point to it.
-func (r *Reconciler) reconcileMachineTemplate(ctx context.Context, tc *butlerv1alpha1.TenantCluster) error {
-	logger := log.FromContext(ctx)
-
-	if tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseReady &&
-		tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseInstalling {
-		return nil
-	}
-	if tc.Status.TenantNamespace == "" {
-		return nil
-	}
-
-	// Get the MachineDeployment
-	mdName := fmt.Sprintf("%s-workers", tc.Name)
-	md := &unstructured.Unstructured{}
-	md.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ClusterAPIGroup,
-		Version: capi.ClusterAPIVersion,
-		Kind:    "MachineDeployment",
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Namespace: tc.Status.TenantNamespace,
-		Name:      mdName,
-	}, md); err != nil {
-		if apierrors.IsNotFound(err) {
-			return nil
-		}
-		return fmt.Errorf("failed to get MachineDeployment: %w", err)
-	}
-
-	// Get current infrastructureRef from MachineDeployment
-	infraRefName, _, _ := unstructured.NestedString(md.Object,
-		"spec", "template", "spec", "infrastructureRef", "name")
-	infraRefKind, _, _ := unstructured.NestedString(md.Object,
-		"spec", "template", "spec", "infrastructureRef", "kind")
-	infraRefAPIVersion, _, _ := unstructured.NestedString(md.Object,
-		"spec", "template", "spec", "infrastructureRef", "apiVersion")
-	if infraRefName == "" || infraRefKind == "" {
-		return nil
-	}
-
-	// Get the current MachineTemplate
-	currentTmpl := &unstructured.Unstructured{}
-	currentTmpl.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.InfrastructureAPIGroup,
-		Version: strings.TrimPrefix(infraRefAPIVersion, capi.InfrastructureAPIGroup+"/"),
-		Kind:    infraRefKind,
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Namespace: tc.Status.TenantNamespace,
-		Name:      infraRefName,
-	}, currentTmpl); err != nil {
-		if apierrors.IsNotFound(err) {
-			return nil
-		}
-		return fmt.Errorf("failed to get current MachineTemplate: %w", err)
-	}
-
-	desiredCPU, desiredMemory, desiredDisk := capi.GetMachineSpecs(tc)
-
-	var currentCPU int64
-	var currentMemory, currentDisk string
-
-	switch infraRefKind {
-	case "KubevirtMachineTemplate":
-		currentCPU, _, _ = unstructured.NestedInt64(currentTmpl.Object,
-			"spec", "template", "spec", "virtualMachineTemplate", "spec",
-			"template", "spec", "domain", "cpu", "cores")
-		currentMemory, _, _ = unstructured.NestedString(currentTmpl.Object,
-			"spec", "template", "spec", "virtualMachineTemplate", "spec",
-			"template", "spec", "domain", "resources", "requests", "memory")
-		dvTemplates, ok, _ := unstructured.NestedSlice(currentTmpl.Object,
-			"spec", "template", "spec", "virtualMachineTemplate", "spec",
-			"dataVolumeTemplates")
-		if ok && len(dvTemplates) > 0 {
-			if dvt, ok := dvTemplates[0].(map[string]interface{}); ok {
-				currentDisk, _, _ = unstructured.NestedString(dvt,
-					"spec", "storage", "resources", "requests", "storage")
-			}
-		}
-
-	case "NutanixMachineTemplate":
-		currentCPU, _, _ = unstructured.NestedInt64(currentTmpl.Object,
-			"spec", "template", "spec", "vcpuSockets")
-		currentMemory, _, _ = unstructured.NestedString(currentTmpl.Object,
-			"spec", "template", "spec", "memorySize")
-		currentDisk, _, _ = unstructured.NestedString(currentTmpl.Object,
-			"spec", "template", "spec", "systemDiskSize")
-
-	default:
-		logger.Info("machine template rolling update not supported for provider",
-			"kind", infraRefKind, "cluster", tc.Name)
-		return nil
-	}
-
-	if currentCPU == desiredCPU && currentMemory == desiredMemory && currentDisk == desiredDisk {
-		return nil
-	}
-
-	logger.Info("machine template change detected",
-		"cluster", tc.Name,
-		"currentCPU", currentCPU, "desiredCPU", desiredCPU,
-		"currentMemory", currentMemory, "desiredMemory", desiredMemory,
-		"currentDisk", currentDisk, "desiredDisk", desiredDisk)
-
-	// Determine new template name — increment version suffix
-	newName := infraRefName + "-v2"
-	// Check if -v2 already exists, if so try -v3, etc.
-	for i := 2; i <= 20; i++ {
-		candidate := fmt.Sprintf("%s-worker-v%d", tc.Name, i)
-		check := &unstructured.Unstructured{}
-		check.SetGroupVersionKind(currentTmpl.GroupVersionKind())
-		err := r.Get(ctx, types.NamespacedName{
-			Namespace: tc.Status.TenantNamespace,
-			Name:      candidate,
-		}, check)
-		if apierrors.IsNotFound(err) {
-			newName = candidate
-			break
-		}
-		if err != nil {
-			return fmt.Errorf("failed to check template name %s: %w", candidate, err)
-		}
-		// Already exists — check if MachineDeployment already points to it
-		if infraRefName == candidate {
-			// Already using this one, need next version
-			continue
-		}
-	}
-
-	// Deep copy the current template and update specs
-	newTmpl := currentTmpl.DeepCopy()
-	newTmpl.SetName(newName)
-	newTmpl.SetResourceVersion("")
-	newTmpl.SetUID("")
-	newTmpl.SetCreationTimestamp(metav1.Time{})
-	// Remove managedFields to avoid conflicts
-	newTmpl.SetManagedFields(nil)
-
-	var setErrs []error
-	setField := func(obj map[string]interface{}, val interface{}, fields ...string) {
-		if err := unstructured.SetNestedField(obj, val, fields...); err != nil {
-			setErrs = append(setErrs, err)
-		}
-	}
-
-	switch infraRefKind {
-	case "KubevirtMachineTemplate":
-		vmBase := []string{"spec", "template", "spec", "virtualMachineTemplate", "spec", "template", "spec", "domain"}
-		setField(newTmpl.Object, desiredCPU, append(vmBase, "cpu", "cores")...)
-		setField(newTmpl.Object, desiredMemory, append(vmBase, "resources", "requests", "memory")...)
-		setField(newTmpl.Object, desiredMemory, append(vmBase, "resources", "limits", "memory")...)
-		setField(newTmpl.Object, fmt.Sprintf("%d", desiredCPU), append(vmBase, "resources", "limits", "cpu")...)
-
-		dvTemplates, ok, _ := unstructured.NestedSlice(newTmpl.Object,
-			"spec", "template", "spec", "virtualMachineTemplate", "spec", "dataVolumeTemplates")
-		if ok && len(dvTemplates) > 0 {
-			if dvt, ok := dvTemplates[0].(map[string]interface{}); ok {
-				setField(dvt, desiredDisk, "spec", "storage", "resources", "requests", "storage")
-				dvTemplates[0] = dvt
-				setField(newTmpl.Object, dvTemplates,
-					"spec", "template", "spec", "virtualMachineTemplate", "spec", "dataVolumeTemplates")
-			}
-		}
-
-	case "NutanixMachineTemplate":
-		setField(newTmpl.Object, desiredCPU, "spec", "template", "spec", "vcpuSockets")
-		setField(newTmpl.Object, desiredMemory, "spec", "template", "spec", "memorySize")
-		setField(newTmpl.Object, desiredDisk, "spec", "template", "spec", "systemDiskSize")
-	}
-
-	if len(setErrs) > 0 {
-		return fmt.Errorf("failed to set fields on new MachineTemplate: %v", setErrs[0])
-	}
-
-	// Create the new template
-	if err := r.Create(ctx, newTmpl); err != nil {
-		if apierrors.IsAlreadyExists(err) {
-			// Race condition — already created, just update MD
-		} else {
-			return fmt.Errorf("failed to create new MachineTemplate %s: %w", newName, err)
-		}
-	} else {
-		logger.Info("created new MachineTemplate",
-			"name", newName,
-			"cpu", desiredCPU,
-			"memory", desiredMemory)
-	}
-
-	// Update MachineDeployment to point to the new template
-	patch := []byte(fmt.Sprintf(`{"spec":{"template":{"spec":{"infrastructureRef":{"name":"%s"}}}}}`, newName))
-	if err := r.Patch(ctx, md, client.RawPatch(types.MergePatchType, patch)); err != nil {
-		return fmt.Errorf("failed to update MachineDeployment infrastructureRef: %w", err)
-	}
-
-	logger.Info("updated MachineDeployment to new MachineTemplate",
-		"machineDeployment", mdName,
-		"newTemplate", newName)
-
-	return nil
-}
-
-func (r *Reconciler) handleKamajiHarvesterCompatibility(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) (bool, error) {
-	logger := log.FromContext(ctx)
-
-	kcp := &unstructured.Unstructured{}
-	kcp.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ControlPlaneAPIGroup,
-		Version: capi.ControlPlaneAPIVersion,
-		Kind:    "StewardControlPlane",
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Name:      tc.Name,
-		Namespace: tc.Status.TenantNamespace,
-	}, kcp); err != nil {
-		if apierrors.IsNotFound(err) {
-			return false, nil
-		}
-		return false, err
-	}
-
-	conditions, found, _ := unstructured.NestedSlice(kcp.Object, "status", "conditions")
-	if !found {
-		return false, nil
-	}
-
-	tcp := &unstructured.Unstructured{}
-	tcp.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "steward.butlerlabs.dev",
-		Version: "v1alpha1",
-		Kind:    "TenantControlPlane",
-	})
-	if err := r.Get(ctx, types.NamespacedName{
-		Name:      tc.Name,
-		Namespace: tc.Status.TenantNamespace,
-	}, tcp); err != nil {
-		if apierrors.IsNotFound(err) {
-			return false, nil
-		}
-		return false, err
-	}
-
-	// Determine the control plane endpoint based on exposure mode
-	var endpointHost string
-	var endpointPort int64
-
-	exposureMode := butlerv1alpha1.ControlPlaneExposureModeLoadBalancer
-	if butlerConfig != nil {
-		exposureMode = butlerConfig.GetControlPlaneExposureMode()
-	}
-
-	switch exposureMode {
-	case butlerv1alpha1.ControlPlaneExposureModeIngress, butlerv1alpha1.ControlPlaneExposureModeGateway:
-		// For Ingress/Gateway modes, use the generated hostname
-		// Ingress uses port 443 (HTTPS), Gateway uses port 6443 (kube-apiserver listener)
-		if butlerConfig != nil {
-			hostnamePattern := butlerConfig.GetControlPlaneExposureHostname()
-			if hostnamePattern != "" {
-				// Generate tenant-specific hostname: "clustername.namespace.k8s.example.com" from "*.k8s.example.com"
-				base := strings.TrimPrefix(hostnamePattern, "*.")
-				endpointHost = fmt.Sprintf("%s.%s.%s", tc.Name, tc.Status.TenantNamespace, base)
-				if exposureMode == butlerv1alpha1.ControlPlaneExposureModeGateway {
-					endpointPort = 6443
-				} else {
-					endpointPort = 443
-				}
-				logger.V(1).Info("using Ingress/Gateway hostname for controlPlaneEndpoint",
-					"mode", exposureMode, "hostname", endpointHost)
-			}
-		}
-	default:
-		// LoadBalancer mode: get IP from service status, port 6443
-		serviceStatus, found, _ := unstructured.NestedMap(tcp.Object, "status", "kubernetesResources", "service")
-		if found {
-			ingress, found, _ := unstructured.NestedSlice(serviceStatus, "loadBalancer", "ingress")
-			if found && len(ingress) > 0 {
-				if ingressEntry, ok := ingress[0].(map[string]interface{}); ok {
-					endpointHost, _, _ = unstructured.NestedString(ingressEntry, "ip")
-					endpointPort = 6443
-					logger.V(1).Info("using LoadBalancer IP for controlPlaneEndpoint",
-						"ip", endpointHost)
-				}
-			}
-		}
-	}
-
-	// Patch Cluster controlPlaneEndpoint if we have a valid endpoint
-	if endpointHost != "" {
-		cluster := &unstructured.Unstructured{}
-		cluster.SetGroupVersionKind(schema.GroupVersionKind{
-			Group:   capi.ClusterAPIGroup,
-			Version: capi.ClusterAPIVersion,
-			Kind:    "Cluster",
-		})
-		if err := r.Get(ctx, types.NamespacedName{
-			Name:      tc.Name,
-			Namespace: tc.Status.TenantNamespace,
-		}, cluster); err == nil {
-			currentHost, _, _ := unstructured.NestedString(cluster.Object, "spec", "controlPlaneEndpoint", "host")
-			currentPort, _, _ := unstructured.NestedInt64(cluster.Object, "spec", "controlPlaneEndpoint", "port")
-			if currentHost != endpointHost || currentPort != endpointPort {
-				logger.Info("patching Cluster controlPlaneEndpoint", "currentHost", currentHost, "newHost", endpointHost, "currentPort", currentPort, "newPort", endpointPort)
-
-				patch := []byte(fmt.Sprintf(
-					`{"spec":{"controlPlaneEndpoint":{"host":%q,"port":%d}}}`,
-					endpointHost, endpointPort,
-				))
-				if err := r.Patch(ctx, cluster, client.RawPatch(types.MergePatchType, patch)); err != nil {
-					logger.Error(err, "failed to patch Cluster controlPlaneEndpoint")
-				} else {
-					logger.Info("successfully patched Cluster controlPlaneEndpoint", "host", endpointHost, "port", endpointPort)
-					return true, nil
-				}
-			}
-		}
-	}
-
-	initialized, _, _ := unstructured.NestedBool(kcp.Object, "status", "initialized")
-	ready, _, _ := unstructured.NestedBool(kcp.Object, "status", "ready")
-	updatedReplicas, _, _ := unstructured.NestedInt64(kcp.Object, "status", "updatedReplicas")
-	version, _, _ := unstructured.NestedString(kcp.Object, "status", "version")
-	if initialized && ready && updatedReplicas >= 1 && version != "" {
-		return false, nil
-	}
-
-	hasUnsupportedProviderError := false
-	for _, c := range conditions {
-		cond, ok := c.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		condType, _, _ := unstructured.NestedString(cond, "type")
-		message, _, _ := unstructured.NestedString(cond, "message")
-		if condType == "InfrastructureClusterPatched" && message == "unsupported infrastructure provider" {
-			hasUnsupportedProviderError = true
-			break
-		}
-	}
-
-	if !hasUnsupportedProviderError {
-		return false, nil
-	}
-
-	logger.Info("detected Kamaji unsupported infrastructure provider error, patching StewardControlPlane status")
-
-	deploymentStatus, deployFound, _ := unstructured.NestedMap(tcp.Object, "status", "kubernetesResources", "deployment")
-	if !deployFound {
-		logger.V(1).Info("TenantControlPlane deployment status not found")
-		return false, nil
-	}
-
-	readyReplicas, _, _ := unstructured.NestedInt64(deploymentStatus, "readyReplicas")
-	if readyReplicas < 1 {
-		logger.V(1).Info("TenantControlPlane deployment not ready yet", "readyReplicas", readyReplicas)
-		return false, nil
-	}
-
-	logger.Info("TenantControlPlane is ready, patching StewardControlPlane status", "readyReplicas", readyReplicas)
-
-	if err := unstructured.SetNestedField(kcp.Object, true, "status", "initialized"); err != nil {
-		return false, fmt.Errorf("failed to set initialized: %w", err)
-	}
-	if err := unstructured.SetNestedField(kcp.Object, true, "status", "ready"); err != nil {
-		return false, fmt.Errorf("failed to set ready: %w", err)
-	}
-	if err := unstructured.SetNestedField(kcp.Object, int64(1), "status", "replicas"); err != nil {
-		return false, fmt.Errorf("failed to set replicas: %w", err)
-	}
-	if err := unstructured.SetNestedField(kcp.Object, int64(1), "status", "readyReplicas"); err != nil {
-		return false, fmt.Errorf("failed to set readyReplicas: %w", err)
-	}
-	if err := unstructured.SetNestedField(kcp.Object, int64(1), "status", "updatedReplicas"); err != nil {
-		return false, fmt.Errorf("failed to set updatedReplicas: %w", err)
-	}
-	if err := unstructured.SetNestedField(kcp.Object, tc.Spec.KubernetesVersion, "status", "version"); err != nil {
-		return false, fmt.Errorf("failed to set version: %w", err)
-	}
-
-	if err := r.Status().Update(ctx, kcp); err != nil {
-		return false, fmt.Errorf("failed to update StewardControlPlane status: %w", err)
-	}
-
-	return true, nil
-}
-
-func (r *Reconciler) getButlerConfig(ctx context.Context) (*butlerv1alpha1.ButlerConfig, error) {
-	config := &butlerv1alpha1.ButlerConfig{}
-	if err := r.Get(ctx, types.NamespacedName{Name: ButlerConfigSingletonName}, config); err != nil {
-		if apierrors.IsNotFound(err) {
-			return &butlerv1alpha1.ButlerConfig{
-				Spec: butlerv1alpha1.ButlerConfigSpec{
-					MultiTenancy: butlerv1alpha1.MultiTenancyConfig{
-						Mode: butlerv1alpha1.MultiTenancyModeDisabled,
-					},
-					DefaultNamespace: "butler-tenants",
-				},
-			}, nil
-		}
-		return nil, err
-	}
-	return config, nil
-}
-
-func (r *Reconciler) validateTenantCluster(ctx context.Context, tc *butlerv1alpha1.TenantCluster, config *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-	mode := config.Spec.MultiTenancy.Mode
-
-	logger.V(1).Info("validating TenantCluster", "mode", mode)
-
-	switch mode {
-	case butlerv1alpha1.MultiTenancyModeEnforced:
-		return r.validateEnforcedMode(ctx, tc)
-	case butlerv1alpha1.MultiTenancyModeOptional:
-		return r.validateOptionalMode(ctx, tc, config)
-	case butlerv1alpha1.MultiTenancyModeDisabled:
-		return r.validateDisabledMode(ctx, tc, config)
-	default:
-		return r.validateDisabledMode(ctx, tc, config)
-	}
-}
-
-func (r *Reconciler) validateEnforcedMode(ctx context.Context, tc *butlerv1alpha1.TenantCluster) error {
-	logger := log.FromContext(ctx)
-
-	if tc.Spec.TeamRef == nil || tc.Spec.TeamRef.Name == "" {
-		return fmt.Errorf("teamRef is required when multi-tenancy mode is Enforced")
-	}
-
-	team := &butlerv1alpha1.Team{}
-	if err := r.Get(ctx, types.NamespacedName{Name: tc.Spec.TeamRef.Name}, team); err != nil {
-		if apierrors.IsNotFound(err) {
-			return fmt.Errorf("team %q not found", tc.Spec.TeamRef.Name)
-		}
-		return fmt.Errorf("failed to get team: %w", err)
-	}
-
-	if team.Status.Phase != butlerv1alpha1.TeamPhaseReady {
-		return fmt.Errorf("team %q is not ready (phase: %s)", team.Name, team.Status.Phase)
-	}
-
-	expectedNamespace := team.Status.Namespace
-	if expectedNamespace == "" {
-		expectedNamespace = team.Name
-	}
-
-	if tc.Namespace != expectedNamespace {
-		return fmt.Errorf("TenantCluster must be in team namespace %q, got %q", expectedNamespace, tc.Namespace)
-	}
-
-	logger.V(1).Info("enforced mode validation passed", "team", team.Name)
-	return nil
-}
-
-func (r *Reconciler) validateOptionalMode(ctx context.Context, tc *butlerv1alpha1.TenantCluster, config *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-
-	if tc.Spec.TeamRef != nil && tc.Spec.TeamRef.Name != "" {
-		team := &butlerv1alpha1.Team{}
-		if err := r.Get(ctx, types.NamespacedName{Name: tc.Spec.TeamRef.Name}, team); err != nil {
-			if apierrors.IsNotFound(err) {
-				return fmt.Errorf("team %q not found", tc.Spec.TeamRef.Name)
-			}
-			return fmt.Errorf("failed to get team: %w", err)
-		}
-
-		expectedNamespace := team.Status.Namespace
-		if expectedNamespace == "" {
-			expectedNamespace = team.Name
-		}
-
-		if tc.Namespace != expectedNamespace {
-			return fmt.Errorf("TenantCluster must be in team namespace %q when teamRef is set, got %q", expectedNamespace, tc.Namespace)
-		}
-
-		logger.V(1).Info("optional mode validation passed with team", "team", team.Name)
-	} else {
-		defaultNS := config.Spec.DefaultNamespace
-		if defaultNS == "" {
-			defaultNS = "butler-tenants"
-		}
-
-		if tc.Namespace != defaultNS {
-			return fmt.Errorf("TenantCluster without teamRef must be in default namespace %q, got %q", defaultNS, tc.Namespace)
-		}
-
-		logger.V(1).Info("optional mode validation passed without team", "namespace", tc.Namespace)
-	}
-
-	return nil
-}
-
-func (r *Reconciler) validateDisabledMode(ctx context.Context, tc *butlerv1alpha1.TenantCluster, config *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-
-	defaultNS := config.Spec.DefaultNamespace
-	if defaultNS == "" {
-		defaultNS = "butler-tenants"
-	}
-
-	if tc.Namespace != defaultNS {
-		return fmt.Errorf("TenantCluster must be in default namespace %q when multi-tenancy is disabled, got %q", defaultNS, tc.Namespace)
-	}
-
-	logger.V(1).Info("disabled mode validation passed", "namespace", tc.Namespace)
-	return nil
-}
-
-func (r *Reconciler) reconcileTenantNamespace(ctx context.Context, tc *butlerv1alpha1.TenantCluster, config *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-
-	if tc.Status.TenantNamespace == "" {
-		tc.Status.TenantNamespace = generateTenantNamespace(tc)
-	}
-
-	nsName := tc.Status.TenantNamespace
-	logger.Info("reconciling tenant namespace", "namespace", nsName)
-
-	ns := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{Name: nsName},
-	}
-
-	op, err := controllerutil.CreateOrUpdate(ctx, r.Client, ns, func() error {
-		if ns.Labels == nil {
-			ns.Labels = make(map[string]string)
-		}
-		ns.Labels[butlerv1alpha1.LabelManagedBy] = "butler"
-		ns.Labels[butlerv1alpha1.LabelTenant] = tc.Name
-		ns.Labels[butlerv1alpha1.LabelSourceNamespace] = tc.Namespace
-		ns.Labels[butlerv1alpha1.LabelSourceName] = tc.Name
-
-		if tc.Spec.TeamRef != nil && tc.Spec.TeamRef.Name != "" {
-			ns.Labels[butlerv1alpha1.LabelTeam] = tc.Spec.TeamRef.Name
-		}
-
-		ns.Labels["pod-security.kubernetes.io/enforce"] = "privileged"
-		ns.Labels["pod-security.kubernetes.io/audit"] = "privileged"
-		ns.Labels["pod-security.kubernetes.io/warn"] = "privileged"
-
-		if ns.Annotations == nil {
-			ns.Annotations = make(map[string]string)
-		}
-		ns.Annotations[butlerv1alpha1.AnnotationDescription] = fmt.Sprintf("Tenant namespace for TenantCluster %s/%s", tc.Namespace, tc.Name)
-
-		return nil
-	})
-
-	if err != nil {
-		return fmt.Errorf("failed to create/update tenant namespace: %w", err)
-	}
-
-	if op != controllerutil.OperationResultNone {
-		logger.Info("tenant namespace reconciled", "namespace", nsName, "operation", op)
-	}
-
-	if tc.Spec.TeamRef != nil && tc.Spec.TeamRef.Name != "" {
-		if err := r.reconcileTeamRoleBinding(ctx, tc, nsName); err != nil {
-			return fmt.Errorf("failed to create team RoleBinding: %w", err)
-		}
-	}
-
-	return nil
-}
-
-func (r *Reconciler) reconcileTeamRoleBinding(ctx context.Context, tc *butlerv1alpha1.TenantCluster, nsName string) error {
-	logger := log.FromContext(ctx)
-
-	team := &butlerv1alpha1.Team{}
-	if err := r.Get(ctx, types.NamespacedName{Name: tc.Spec.TeamRef.Name}, team); err != nil {
-		return err
-	}
-
-	rbName := fmt.Sprintf("butler-team-%s-access", team.Name)
-
-	rb := &rbacv1.RoleBinding{
-		ObjectMeta: metav1.ObjectMeta{Name: rbName, Namespace: nsName},
-	}
-
-	op, err := controllerutil.CreateOrUpdate(ctx, r.Client, rb, func() error {
-		if rb.Labels == nil {
-			rb.Labels = make(map[string]string)
-		}
-		rb.Labels[butlerv1alpha1.LabelManagedBy] = "butler"
-		rb.Labels[butlerv1alpha1.LabelTeam] = team.Name
-		rb.Labels[butlerv1alpha1.LabelTenant] = tc.Name
-
-		rb.RoleRef = rbacv1.RoleRef{
-			APIGroup: rbacv1.GroupName,
-			Kind:     "ClusterRole",
-			Name:     "admin",
-		}
-
-		var subjects []rbacv1.Subject
-		for _, user := range team.Spec.Access.Users {
-			subjects = append(subjects, rbacv1.Subject{
-				Kind: rbacv1.UserKind, APIGroup: rbacv1.GroupName, Name: user.Name,
-			})
-		}
-		for _, group := range team.Spec.Access.Groups {
-			subjects = append(subjects, rbacv1.Subject{
-				Kind: rbacv1.GroupKind, APIGroup: rbacv1.GroupName, Name: group.Name,
-			})
-		}
-		rb.Subjects = subjects
-		return nil
-	})
-
-	if err != nil {
-		return err
-	}
-
-	if op != controllerutil.OperationResultNone {
-		logger.Info("team RoleBinding reconciled", "name", rbName, "namespace", nsName, "operation", op)
-	}
-
-	return nil
-}
-
-func (r *Reconciler) handleDeletion(ctx context.Context, tc *butlerv1alpha1.TenantCluster) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-	logger.Info("handling TenantCluster deletion", "name", tc.Name, "namespace", tc.Namespace)
-
-	tc.Status.Phase = butlerv1alpha1.TenantClusterPhaseDeleting
-	r.Recorder.Eventf(tc, corev1.EventTypeNormal, "ClusterDeleting", "Cluster %s deletion started", tc.Name)
-	if err := r.Status().Update(ctx, tc); err != nil {
-		if !apierrors.IsConflict(err) {
-			return ctrl.Result{}, err
-		}
-	}
-
-	// Evict cached tenant client before cleanup
-	if r.ClientManager != nil && tc.Status.TenantNamespace != "" {
-		r.ClientManager.RemoveClient(tc.Status.TenantNamespace, tc.Name)
-	}
-
-	// Clean up IPAllocations before deleting namespace
-	r.cleanupIPAllocations(ctx, tc)
-
-	if tc.Status.TenantNamespace != "" {
-		// Delete the CAPI Cluster first so the CAPI controller can remove its
-		// finalizer while dependent resources (KubevirtCluster, StewardControlPlane)
-		// still exist. Deleting the namespace first cascade-deletes those resources,
-		// leaving the CAPI Cluster with a finalizer that can never be removed.
-		capiCluster := &unstructured.Unstructured{}
-		capiCluster.SetGroupVersionKind(schema.GroupVersionKind{
-			Group:   capi.ClusterAPIGroup,
-			Version: capi.ClusterAPIVersion,
-			Kind:    "Cluster",
-		})
-		if err := r.Get(ctx, types.NamespacedName{
-			Name:      tc.Name,
-			Namespace: tc.Status.TenantNamespace,
-		}, capiCluster); err == nil {
-			if capiCluster.GetDeletionTimestamp().IsZero() {
-				logger.Info("deleting CAPI Cluster before namespace", "cluster", tc.Name)
-				if err := r.Delete(ctx, capiCluster); err != nil && !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, err
-				}
-			}
-			// CAPI Cluster still exists (finalizer pending), wait for CAPI controller to clean up
-			logger.V(1).Info("waiting for CAPI Cluster deletion", "cluster", tc.Name)
-			return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
-		} else if !apierrors.IsNotFound(err) {
-			return ctrl.Result{}, err
-		}
-
-		// CAPI Cluster is gone, safe to delete the namespace
-		ns := &corev1.Namespace{}
-		err := r.Get(ctx, types.NamespacedName{Name: tc.Status.TenantNamespace}, ns)
-		if err == nil {
-			logger.Info("deleting tenant namespace", "namespace", tc.Status.TenantNamespace)
-			if err := r.Delete(ctx, ns); err != nil && !apierrors.IsNotFound(err) {
-				return ctrl.Result{}, err
-			}
-			return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
-		} else if !apierrors.IsNotFound(err) {
-			return ctrl.Result{}, err
-		}
-		logger.Info("tenant namespace deleted", "namespace", tc.Status.TenantNamespace)
-	}
-
-	latest := &butlerv1alpha1.TenantCluster{}
-	if err := r.Get(ctx, client.ObjectKeyFromObject(tc), latest); err != nil {
-		if apierrors.IsNotFound(err) {
-			return ctrl.Result{}, nil
-		}
-		return ctrl.Result{}, err
-	}
-
-	if controllerutil.ContainsFinalizer(latest, butlerv1alpha1.FinalizerTenantCluster) {
-		controllerutil.RemoveFinalizer(latest, butlerv1alpha1.FinalizerTenantCluster)
-		if err := r.Update(ctx, latest); err != nil {
-			if apierrors.IsNotFound(err) {
-				return ctrl.Result{}, nil
-			}
-			if apierrors.IsConflict(err) {
-				return ctrl.Result{Requeue: true}, nil
-			}
-			return ctrl.Result{}, err
-		}
-	}
-
-	logger.Info("TenantCluster deletion complete", "name", tc.Name)
-	return ctrl.Result{}, nil
-}
-
-// reconcileImageSync ensures the OS image is synced to the infrastructure provider
-// via the Butler Image Factory. It creates an ImageSync resource if needed, deduplicates
-// across TenantClusters sharing the same schematic+version+provider, and returns the
-// provider-specific image reference once the sync is complete.
-//
-// Returns:
-//   - providerImageRef: the provider-specific image reference (e.g., "default/talos-v1.12.4-amd64")
-//     Empty string if no sync is needed (no schematic configured or auto-sync disabled).
-//   - error: non-nil triggers a requeue (image sync in progress, failed, or creation error).
-func (r *Reconciler) reconcileImageSync(ctx context.Context, tc *butlerv1alpha1.TenantCluster, pc *butlerv1alpha1.ProviderConfig, bc *butlerv1alpha1.ButlerConfig) (string, error) {
-	logger := log.FromContext(ctx)
-
-	// Resolve schematicID: TC spec takes precedence, then ButlerConfig default
-	schematicID := tc.Spec.Workers.MachineTemplate.OS.SchematicID
-	if schematicID == "" && bc != nil && bc.Spec.ImageFactory != nil {
-		schematicID = bc.Spec.ImageFactory.DefaultSchematicID
-	}
-	if schematicID == "" {
-		// No schematic configured — no image sync needed
-		return "", nil
-	}
-
-	// Image factory must be configured if a schematicID is set
-	if bc == nil || !bc.IsImageFactoryConfigured() {
-		return "", fmt.Errorf("schematicID %q is set but Image Factory is not configured in ButlerConfig", schematicID)
-	}
-
-	// Check if auto-sync is enabled
-	if !bc.IsAutoSyncEnabled() {
-		logger.V(1).Info("image auto-sync disabled, skipping ImageSync creation")
-		return "", nil
-	}
-
-	// Resolve image version, architecture, and platform
-	version := tc.Spec.Workers.MachineTemplate.OS.Version
-	if version == "" {
-		version = "9.5" // default OS version from kubebuilder marker
-	}
-	// Architecture defaults to amd64 — all current Butler providers target amd64.
-	// When multi-arch support is added, this should read from MachineTemplate or OS spec.
-	arch := "amd64"
-	// Platform is the artifact name prefix in the factory URL.
-	// For Butler Image Factory: use the OS type (talos, flatcar, kairos, bottlerocket).
-	// For Siderolabs factory (factory.talos.dev): use "nocloud" (or metal, vmware, etc.).
-	platform := string(tc.Spec.Workers.MachineTemplate.OS.Type)
-	if platform == "" {
-		platform = "nocloud"
-	}
-
-	// Truncate schematicID to 63 chars for label value (Kubernetes label limit)
-	labelSchematicID := schematicID
-	if len(labelSchematicID) > 63 {
-		labelSchematicID = labelSchematicID[:63]
-	}
-
-	// Build label selector for deduplication
-	matchLabels := map[string]string{
-		butlerv1alpha1.LabelSchematicID:    labelSchematicID,
-		butlerv1alpha1.LabelImageVersion:   version,
-		butlerv1alpha1.LabelProviderConfig: pc.Name,
-		butlerv1alpha1.LabelImageArch:      arch,
-	}
-
-	// List existing ImageSyncs matching labels in the TC's namespace
-	var existingList butlerv1alpha1.ImageSyncList
-	if err := r.List(ctx, &existingList,
-		client.InNamespace(tc.Namespace),
-		client.MatchingLabels(matchLabels),
-	); err != nil {
-		return "", fmt.Errorf("failed to list ImageSyncs: %w", err)
-	}
-
-	if len(existingList.Items) > 0 {
-		is := &existingList.Items[0]
-
-		switch {
-		case is.IsReady():
-			logger.V(1).Info("ImageSync is ready", "name", is.Name, "providerImageRef", is.Status.ProviderImageRef)
-			r.setCondition(tc, butlerv1alpha1.TenantClusterConditionImageReady,
-				metav1.ConditionTrue, "ImageReady", "OS image synced to provider")
-			tc.Status.ImageSyncRef = &butlerv1alpha1.LocalObjectReference{Name: is.Name}
-			return is.Status.ProviderImageRef, nil
-
-		case is.IsFailed():
-			reason := is.Status.FailureReason
-			if reason == "" {
-				reason = "ImageSyncFailed"
-			}
-			message := is.Status.FailureMessage
-			if message == "" {
-				message = "Image sync failed"
-			}
-			r.setCondition(tc, butlerv1alpha1.TenantClusterConditionImageReady,
-				metav1.ConditionFalse, reason, message)
-			tc.Status.ImageSyncRef = &butlerv1alpha1.LocalObjectReference{Name: is.Name}
-			return "", fmt.Errorf("ImageSync %s failed: %s", is.Name, message)
-
-		default:
-			// In progress (Pending, Building, Downloading, Uploading)
-			logger.Info("ImageSync in progress", "name", is.Name, "phase", is.Status.Phase)
-			r.setCondition(tc, butlerv1alpha1.TenantClusterConditionImageReady,
-				metav1.ConditionFalse, "ImageSyncInProgress",
-				fmt.Sprintf("Image sync %s is %s", is.Name, is.Status.Phase))
-			tc.Status.ImageSyncRef = &butlerv1alpha1.LocalObjectReference{Name: is.Name}
-			return "", fmt.Errorf("ImageSync %s is in progress (phase: %s)", is.Name, is.Status.Phase)
-		}
-	}
-
-	// No existing ImageSync found — create one
-
-	// Build a short, DNS-safe name
-	schematicPrefix := schematicID
-	if len(schematicPrefix) > 8 {
-		schematicPrefix = schematicPrefix[:8]
-	}
-	// Sanitize version for DNS label (replace dots with dashes, strip leading 'v')
-	sanitizedVersion := strings.ReplaceAll(version, ".", "-")
-	sanitizedVersion = strings.TrimPrefix(sanitizedVersion, "v")
-	imageSyncName := fmt.Sprintf("%s-%s-%s", tc.Name, schematicPrefix, sanitizedVersion)
-	// Ensure name is DNS-safe (max 253 chars for object names, but keep it reasonable)
-	if len(imageSyncName) > 63 {
-		imageSyncName = imageSyncName[:63]
-	}
-
-	// Resolve provider config reference
-	pcRef := butlerv1alpha1.ProviderReference{
-		Name: pc.Name,
-	}
-	if pc.Namespace != "" && pc.Namespace != tc.Namespace {
-		pcRef.Namespace = pc.Namespace
-	}
-
-	is := &butlerv1alpha1.ImageSync{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      imageSyncName,
-			Namespace: tc.Namespace,
-			Labels: map[string]string{
-				butlerv1alpha1.LabelManagedBy:       "butler",
-				butlerv1alpha1.LabelTenant:          tc.Name,
-				butlerv1alpha1.LabelSchematicID:     labelSchematicID,
-				butlerv1alpha1.LabelImageVersion:    version,
-				butlerv1alpha1.LabelProviderConfig:  pc.Name,
-				butlerv1alpha1.LabelImageArch:       arch,
-			},
-		},
-		Spec: butlerv1alpha1.ImageSyncSpec{
-			FactoryRef: butlerv1alpha1.ImageFactoryRef{
-				SchematicID: schematicID,
-				Version:     version,
-				Arch:        arch,
-				Platform:    platform,
-			},
-			ProviderConfigRef: pcRef,
-			Format:            "qcow2",
-			TransferMode:      butlerv1alpha1.TransferModeDirect,
-		},
-	}
-
-	// Set owner reference for garbage collection
-	if err := controllerutil.SetControllerReference(tc, is, r.Scheme); err != nil {
-		return "", fmt.Errorf("failed to set owner reference on ImageSync: %w", err)
-	}
-
-	logger.Info("creating ImageSync", "name", is.Name, "schematicID", schematicID, "version", version)
-	if err := r.Create(ctx, is); err != nil {
-		if apierrors.IsAlreadyExists(err) {
-			// Race condition — another reconcile created it. Requeue to pick it up.
-			return "", fmt.Errorf("ImageSync %s was just created, requeueing", imageSyncName)
-		}
-		return "", fmt.Errorf("failed to create ImageSync %s: %w", imageSyncName, err)
-	}
-
-	r.setCondition(tc, butlerv1alpha1.TenantClusterConditionImageReady,
-		metav1.ConditionFalse, "ImageSyncPending", "Image sync created, waiting for provider to process")
-	tc.Status.ImageSyncRef = &butlerv1alpha1.LocalObjectReference{Name: is.Name}
-
-	return "", fmt.Errorf("ImageSync %s created, waiting for completion", imageSyncName)
-}
-
 func (r *Reconciler) setFailedStatus(ctx context.Context, tc *butlerv1alpha1.TenantCluster, reason, message string) (ctrl.Result, error) {
 	tc.Status.Phase = butlerv1alpha1.TenantClusterPhaseFailed
 	tc.Status.ObservedGeneration = tc.Generation
@@ -2372,1186 +343,6 @@ func (r *Reconciler) calculateRequeueInterval(tc *butlerv1alpha1.TenantCluster) 
 		return 5 * time.Minute
 	default:
 		return 15 * time.Minute
-	}
-}
-
-func generateTenantNamespace(tc *butlerv1alpha1.TenantCluster) string {
-	uid := string(tc.UID)
-	if len(uid) > 8 {
-		uid = uid[:8]
-	}
-	return fmt.Sprintf("%s-%s", tc.Name, uid)
-}
-
-func extractAPIServerEndpoint(kubeconfig []byte) (string, string) {
-	// Parse kubeconfig to get server URL
-	var kc struct {
-		Clusters []struct {
-			Cluster struct {
-				Server string `yaml:"server"`
-			} `yaml:"cluster"`
-		} `yaml:"clusters"`
-	}
-	if err := yaml.Unmarshal(kubeconfig, &kc); err != nil || len(kc.Clusters) == 0 {
-		return "kubernetes.default.svc.cluster.local", "443" // fallback
-	}
-
-	// Parse URL like https://10.40.0.111:6443
-	server := kc.Clusters[0].Cluster.Server
-	server = strings.TrimPrefix(server, "https://")
-	server = strings.TrimPrefix(server, "http://")
-
-	parts := strings.Split(server, ":")
-	if len(parts) == 2 {
-		return parts[0], parts[1]
-	}
-	return parts[0], "6443"
-}
-
-// getIngressControllerIP returns the external IP of the Ingress controller service.
-// For Ingress/Gateway modes, worker VMs need this IP to resolve the API server hostname
-// before DNS is available (via /etc/hosts entry).
-func (r *Reconciler) getIngressControllerIP(ctx context.Context, butlerConfig *butlerv1alpha1.ButlerConfig) (string, error) {
-	logger := log.FromContext(ctx)
-
-	// Determine the controller type to find the right service
-	controllerType := "traefik" // default
-	if butlerConfig != nil && butlerConfig.Spec.ControlPlaneExposure != nil {
-		if ct := butlerConfig.Spec.ControlPlaneExposure.ControllerType; ct != "" {
-			controllerType = ct
-		}
-	}
-
-	// Map controller type to namespace and service name patterns
-	// Each ingress controller has different naming conventions
-	var namespace, serviceName string
-	switch controllerType {
-	case "traefik":
-		namespace = "traefik"
-		serviceName = "traefik"
-	case "nginx":
-		namespace = "ingress-nginx"
-		serviceName = "ingress-nginx-controller"
-	case "haproxy":
-		namespace = "haproxy-controller"
-		serviceName = "haproxy-kubernetes-ingress"
-	default:
-		// For generic or unknown types, try traefik as fallback
-		namespace = "traefik"
-		serviceName = "traefik"
-	}
-
-	// Look up the service
-	svc := &corev1.Service{}
-	if err := r.Get(ctx, types.NamespacedName{Name: serviceName, Namespace: namespace}, svc); err != nil {
-		if apierrors.IsNotFound(err) {
-			logger.V(1).Info("ingress controller service not found",
-				"namespace", namespace, "service", serviceName, "controllerType", controllerType)
-			return "", nil
-		}
-		return "", fmt.Errorf("failed to get ingress controller service: %w", err)
-	}
-
-	// Extract external IP from LoadBalancer status
-	if svc.Spec.Type != corev1.ServiceTypeLoadBalancer {
-		logger.V(1).Info("ingress controller service is not LoadBalancer type",
-			"type", svc.Spec.Type, "namespace", namespace, "service", serviceName)
-		return "", nil
-	}
-
-	for _, ingress := range svc.Status.LoadBalancer.Ingress {
-		if ingress.IP != "" {
-			logger.V(1).Info("found ingress controller IP",
-				"ip", ingress.IP, "namespace", namespace, "service", serviceName)
-			return ingress.IP, nil
-		}
-		if ingress.Hostname != "" {
-			// Some cloud providers use hostname instead of IP
-			logger.V(1).Info("found ingress controller hostname (no IP)",
-				"hostname", ingress.Hostname, "namespace", namespace, "service", serviceName)
-			return ingress.Hostname, nil
-		}
-	}
-
-	logger.V(1).Info("ingress controller service has no external IP yet",
-		"namespace", namespace, "service", serviceName)
-	return "", nil
-}
-
-// getGatewayIP retrieves the external IP from a Gateway resource status.
-// For Gateway API exposure mode, the external IP is on the Gateway resource
-// rather than an Ingress controller Service.
-func (r *Reconciler) getGatewayIP(ctx context.Context, butlerConfig *butlerv1alpha1.ButlerConfig) (string, error) {
-	logger := log.FromContext(ctx)
-
-	gatewayRef := butlerConfig.GetControlPlaneExposureGatewayRef()
-	if gatewayRef == "" {
-		return "", fmt.Errorf("Gateway mode requires gatewayRef in ButlerConfig")
-	}
-
-	parts := strings.Split(gatewayRef, "/")
-	if len(parts) != 2 {
-		return "", fmt.Errorf("invalid gatewayRef format: %s, expected namespace/name", gatewayRef)
-	}
-	namespace, name := parts[0], parts[1]
-
-	gw := &unstructured.Unstructured{}
-	gw.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "gateway.networking.k8s.io",
-		Version: "v1",
-		Kind:    "Gateway",
-	})
-
-	if err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: name}, gw); err != nil {
-		if apierrors.IsNotFound(err) {
-			logger.V(1).Info("Gateway resource not found", "namespace", namespace, "name", name)
-			return "", nil
-		}
-		return "", fmt.Errorf("failed to get Gateway %s/%s: %w", namespace, name, err)
-	}
-
-	addresses, found, err := unstructured.NestedSlice(gw.Object, "status", "addresses")
-	if err != nil || !found || len(addresses) == 0 {
-		logger.V(1).Info("Gateway has no addresses in status yet", "namespace", namespace, "name", name)
-		return "", nil
-	}
-
-	firstAddr, ok := addresses[0].(map[string]interface{})
-	if !ok {
-		return "", fmt.Errorf("invalid address format in Gateway %s/%s status", namespace, name)
-	}
-
-	ip, found, err := unstructured.NestedString(firstAddr, "value")
-	if err != nil || !found || ip == "" {
-		logger.V(1).Info("Gateway has no IP value in address", "namespace", namespace, "name", name)
-		return "", nil
-	}
-
-	logger.V(1).Info("found Gateway IP", "ip", ip, "namespace", namespace, "name", name)
-	return ip, nil
-}
-
-// getExposureIP returns the external IP for control plane exposure based on mode.
-// For Gateway mode, reads the Gateway resource status. For Ingress mode, reads the
-// Ingress controller Service status.
-func (r *Reconciler) getExposureIP(ctx context.Context, butlerConfig *butlerv1alpha1.ButlerConfig) (string, error) {
-	mode := butlerConfig.GetControlPlaneExposureMode()
-	if mode == butlerv1alpha1.ControlPlaneExposureModeGateway {
-		return r.getGatewayIP(ctx, butlerConfig)
-	}
-	return r.getIngressControllerIP(ctx, butlerConfig)
-}
-
-// reconcileTalosApplyConfig applies Talos machine configs to CAPI Machines in maintenance mode.
-// It discovers Machine IPs, checks whether config was already applied (via annotation),
-// and runs talosctl apply-config --insecure for each unconfigured Machine.
-// reconcileTalosApplyConfig applies Talos machine config to CAPI Machines via talosctl.
-// Returns (allApplied, error) where allApplied is true when all Machines with IPs
-// have their config applied (or are already configured).
-func (r *Reconciler) reconcileTalosApplyConfig(ctx context.Context, tc *butlerv1alpha1.TenantCluster) (bool, error) {
-	logger := log.FromContext(ctx)
-
-	tenantNS := tc.Status.TenantNamespace
-	if tenantNS == "" {
-		return false, nil
-	}
-
-	// Read bootstrap Secret to get the machine config
-	secretName := fmt.Sprintf("%s-talos-bootstrap", tc.Name)
-	bootstrapSecret := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{Name: secretName, Namespace: tenantNS}, bootstrapSecret); err != nil {
-		if apierrors.IsNotFound(err) {
-			return false, nil // Bootstrap Secret not ready yet
-		}
-		return false, fmt.Errorf("failed to get bootstrap Secret: %w", err)
-	}
-
-	machineConfig := bootstrapSecret.Data["value"]
-	if len(machineConfig) == 0 {
-		return false, fmt.Errorf("bootstrap Secret %s has empty machine config", secretName)
-	}
-
-	// Get CAPI Machine addresses
-	machines, err := talos.GetMachineAddresses(ctx, r.Client, tc.Name, tenantNS)
-	if err != nil {
-		return false, fmt.Errorf("getting Machine addresses: %w", err)
-	}
-	if len(machines) == 0 {
-		logger.Info("waiting for CAPI Machines to get IP addresses")
-		return false, nil
-	}
-
-	// Apply config to each Machine that hasn't been configured yet
-	talosClient := talos.NewClient()
-	var failCount, pendingCount int
-	for _, machine := range machines {
-		if machine.Annotations != nil && machine.Annotations["butler.butlerlabs.dev/talos-config-applied"] == "true" {
-			continue
-		}
-
-		pendingCount++
-		hostnamePatch := fmt.Sprintf(`{"machine":{"network":{"hostname":%q}}}`, machine.Name)
-		logger.Info("applying Talos config to Machine", "machine", machine.Name, "ip", machine.IP)
-		if err := talosClient.ApplyConfig(ctx, machine.IP, machineConfig, hostnamePatch); err != nil {
-			// "certificate required" means the node left maintenance mode and already has config.
-			// Treat as success and annotate.
-			if strings.Contains(err.Error(), "certificate required") {
-				logger.Info("node already configured (requires TLS client cert), marking as applied",
-					"machine", machine.Name, "ip", machine.IP)
-			} else {
-				logger.Error(err, "failed to apply Talos config", "machine", machine.Name, "ip", machine.IP)
-				failCount++
-				continue
-			}
-		}
-
-		// Annotate the Machine to mark config as applied
-		if err := r.annotateMachine(ctx, machine.Name, tenantNS); err != nil {
-			logger.Error(err, "failed to annotate Machine after config apply", "machine", machine.Name)
-		}
-		pendingCount--
-	}
-
-	if failCount > 0 && failCount == len(machines) {
-		return false, fmt.Errorf("failed to apply Talos config to all %d Machines", failCount)
-	}
-
-	allApplied := pendingCount == 0 && failCount == 0
-	return allApplied, nil
-}
-
-// annotateMachine adds the talos-config-applied annotation to a CAPI Machine.
-func (r *Reconciler) annotateMachine(ctx context.Context, machineName, namespace string) error {
-	machine := &unstructured.Unstructured{}
-	machine.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   capi.ClusterAPIGroup,
-		Version: capi.ClusterAPIVersion,
-		Kind:    "Machine",
-	})
-	if err := r.Get(ctx, types.NamespacedName{Name: machineName, Namespace: namespace}, machine); err != nil {
-		return err
-	}
-
-	patch := []byte(`{"metadata":{"annotations":{"butler.butlerlabs.dev/talos-config-applied":"true"}}}`)
-	return r.Patch(ctx, machine, client.RawPatch(types.MergePatchType, patch))
-}
-
-// reconcileTalosconfig creates a talosconfig Secret for CLI access to worker nodes.
-// This requires admin client credentials from the trustd-creds Secret (generated by steward).
-func (r *Reconciler) reconcileTalosconfig(ctx context.Context, tc *butlerv1alpha1.TenantCluster) error {
-	logger := log.FromContext(ctx)
-
-	tenantNS := tc.Status.TenantNamespace
-	talosconfigName := fmt.Sprintf("%s-talosconfig", tc.Name)
-
-	// Check if talosconfig Secret already exists
-	existing := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{Name: talosconfigName, Namespace: tenantNS}, existing); err == nil {
-		return nil // Already exists
-	}
-
-	// Read trustd credentials
-	trustdCredsName := fmt.Sprintf("%s-trustd-creds", tc.Name)
-	trustdCreds := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{Name: trustdCredsName, Namespace: tenantNS}, trustdCreds); err != nil {
-		return fmt.Errorf("failed to get trustd credentials: %w", err)
-	}
-
-	adminCert := trustdCreds.Data["admin.crt"]
-	adminKey := trustdCreds.Data["admin.key"]
-	caCert := trustdCreds.Data["os-ca.crt"]
-
-	if len(adminCert) == 0 || len(adminKey) == 0 {
-		logger.V(1).Info("admin credentials not available in trustd-creds, skipping talosconfig")
-		return nil // Steward hasn't been updated yet
-	}
-
-	// Determine the endpoint for the talosconfig
-	var endpoint string
-	machines, err := talos.GetMachineAddresses(ctx, r.Client, tc.Name, tenantNS)
-	if err == nil && len(machines) > 0 {
-		endpoint = machines[0].IP
-	}
-	if endpoint == "" {
-		return nil // No machines with IPs yet
-	}
-
-	// Generate talosconfig YAML
-	talosconfigData, err := talos.GenerateTalosconfig(talos.TalosconfigInput{
-		ContextName: tc.Name,
-		Endpoints:   []string{endpoint},
-		CACert:      caCert,
-		AdminCert:   adminCert,
-		AdminKey:    adminKey,
-	})
-	if err != nil {
-		return fmt.Errorf("generating talosconfig: %w", err)
-	}
-
-	// Create the Secret
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      talosconfigName,
-			Namespace: tenantNS,
-			Labels: map[string]string{
-				butlerv1alpha1.LabelManagedBy: "butler",
-				butlerv1alpha1.LabelTenant:    tc.Name,
-			},
-		},
-		Data: map[string][]byte{
-			"talosconfig": talosconfigData,
-		},
-	}
-
-	// Skip owner reference — TC is in a different namespace than the Secret.
-	// Labels track ownership instead.
-
-	if err := r.Create(ctx, secret); err != nil {
-		if apierrors.IsAlreadyExists(err) {
-			return nil
-		}
-		return fmt.Errorf("creating talosconfig Secret: %w", err)
-	}
-
-	logger.Info("created talosconfig Secret", "name", talosconfigName, "namespace", tenantNS)
-
-	return nil
-}
-
-// isTalosCluster returns true if the TenantCluster uses Talos OS for workers.
-func isTalosCluster(tc *butlerv1alpha1.TenantCluster) bool {
-	return talos.IsTalosCluster(tc)
-}
-
-// effectiveServiceType determines the actual Kubernetes Service type for the TCP.
-// Gateway/Ingress modes default to ClusterIP, LoadBalancer mode defaults to LoadBalancer.
-// The TC's spec.controlPlane.serviceType can override any default.
-func (r *Reconciler) effectiveServiceType(tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) string {
-	if tc.Spec.ControlPlane.ServiceType != "" {
-		return tc.Spec.ControlPlane.ServiceType
-	}
-	if butlerConfig != nil {
-		mode := butlerConfig.GetControlPlaneExposureMode()
-		if mode == butlerv1alpha1.ControlPlaneExposureModeIngress ||
-			mode == butlerv1alpha1.ControlPlaneExposureModeGateway {
-			return "ClusterIP"
-		}
-	}
-	return "LoadBalancer"
-}
-
-// reconcileTalosBootstrap generates the Talos machine config and creates the bootstrap
-// Secret that CAPI's dataSecretName mechanism reads. This is called after the TCP is
-// ready and the control plane is accessible.
-//
-// The Secret contains a Talos v1alpha1 worker machine config with:
-//   - machine.token: trustd token for apid auth with steward-trustd
-//   - cluster.token: kubeadm bootstrap token for kubelet TLS bootstrapping
-//   - machine.ca.crt: OS CA cert for trusting steward-trustd
-//   - cluster.ca.crt: K8s cluster CA for trusting the API server
-func (r *Reconciler) reconcileTalosBootstrap(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) error {
-	logger := log.FromContext(ctx)
-
-	secretName := fmt.Sprintf("%s-talos-bootstrap", tc.Name)
-	tenantNS := tc.Status.TenantNamespace
-
-	// Check if bootstrap Secret already exists
-	existing := &corev1.Secret{}
-	err := r.Get(ctx, types.NamespacedName{Name: secretName, Namespace: tenantNS}, existing)
-	if err == nil {
-		logger.V(1).Info("Talos bootstrap Secret already exists", "name", secretName)
-		return nil
-	}
-	if !apierrors.IsNotFound(err) {
-		return fmt.Errorf("failed to check for bootstrap Secret: %w", err)
-	}
-
-	// Read trustd credentials from the management cluster
-	trustdCredsName := fmt.Sprintf("%s-trustd-creds", tc.Name)
-	trustdCreds := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{Name: trustdCredsName, Namespace: tenantNS}, trustdCreds); err != nil {
-		if apierrors.IsNotFound(err) {
-			logger.Info("trustd credentials not ready yet, waiting", "name", trustdCredsName)
-			return nil
-		}
-		return fmt.Errorf("failed to get trustd credentials: %w", err)
-	}
-
-	machineToken := string(trustdCreds.Data["token"])
-	osCACert := string(trustdCreds.Data["os-ca.crt"])
-	if machineToken == "" || osCACert == "" {
-		return fmt.Errorf("trustd credentials missing token or os-ca.crt")
-	}
-
-	// Read cluster CA from the management cluster
-	caSecretName := fmt.Sprintf("%s-ca", tc.Name)
-	caSecret := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{Name: caSecretName, Namespace: tenantNS}, caSecret); err != nil {
-		if apierrors.IsNotFound(err) {
-			logger.Info("cluster CA not ready yet, waiting", "name", caSecretName)
-			return nil
-		}
-		return fmt.Errorf("failed to get cluster CA: %w", err)
-	}
-
-	clusterCACert := string(caSecret.Data["ca.crt"])
-	if clusterCACert == "" {
-		return fmt.Errorf("cluster CA Secret missing ca.crt")
-	}
-
-	// Determine control plane endpoint based on service type.
-	// When using LoadBalancer, use the Service's external IP directly so that
-	// both the API server (port 6443) and trustd (port 50001) are reachable
-	// on the same address. Gateway mode TLS passthrough on multiple ports
-	// for the same hostname has known issues (cilium/cilium#42898).
-	var controlPlaneEndpoint string
-
-	effectiveServiceType := r.effectiveServiceType(tc, butlerConfig)
-	if effectiveServiceType == "LoadBalancer" {
-		svc := &corev1.Service{}
-		if err := r.Get(ctx, types.NamespacedName{Name: tc.Name, Namespace: tenantNS}, svc); err != nil {
-			if apierrors.IsNotFound(err) {
-				logger.Info("TCP Service not found yet, waiting")
-				return nil
-			}
-			return fmt.Errorf("failed to get TCP Service: %w", err)
-		}
-		var lbIP string
-		for _, ingress := range svc.Status.LoadBalancer.Ingress {
-			if ingress.IP != "" {
-				lbIP = ingress.IP
-				break
-			}
-			if ingress.Hostname != "" {
-				lbIP = ingress.Hostname
-				break
-			}
-		}
-		if lbIP == "" {
-			logger.Info("TCP Service LoadBalancer IP not assigned yet, waiting")
-			return nil
-		}
-		controlPlaneEndpoint = fmt.Sprintf("https://%s:6443", lbIP)
-		logger.Info("using LoadBalancer IP for Talos control plane endpoint", "ip", lbIP)
-	} else {
-		// Gateway/Ingress mode — use the TCP status endpoint (hostname-based)
-		tcp := &unstructured.Unstructured{}
-		tcp.SetGroupVersionKind(schema.GroupVersionKind{
-			Group:   "steward.butlerlabs.dev",
-			Version: "v1alpha1",
-			Kind:    "TenantControlPlane",
-		})
-		if err := r.Get(ctx, types.NamespacedName{Name: tc.Name, Namespace: tenantNS}, tcp); err != nil {
-			return fmt.Errorf("failed to get TenantControlPlane: %w", err)
-		}
-		tcpEndpoint, found, _ := unstructured.NestedString(tcp.Object, "status", "controlPlaneEndpoint")
-		if !found || tcpEndpoint == "" {
-			logger.Info("TCP controlPlaneEndpoint not set yet, waiting")
-			return nil
-		}
-		controlPlaneEndpoint = talos.EndpointFromTCPStatus(tcpEndpoint)
-		logger.Info("using Gateway endpoint for Talos control plane endpoint", "endpoint", controlPlaneEndpoint)
-	}
-
-	// Get tenant client to create bootstrap token in tenant API server
-	tenantClient, err := r.getTenantClient(ctx, tc)
-	if err != nil {
-		logger.Info("tenant client not available yet, waiting", "error", err)
-		return nil
-	}
-
-	// Reuse existing bootstrap token if one was created by a previous attempt,
-	// otherwise generate a new one. This ensures idempotency when the reconciler
-	// succeeds in creating the token but fails before creating the bootstrap Secret.
-	bootstrapToken, err := talos.FindExistingBootstrapToken(ctx, tenantClient.Clientset)
-	if err != nil {
-		return fmt.Errorf("failed to check for existing bootstrap token: %w", err)
-	}
-
-	if bootstrapToken != "" {
-		logger.Info("reusing existing kubeadm bootstrap token in tenant API server")
-	} else {
-		bootstrapToken, err = talos.GenerateBootstrapToken()
-		if err != nil {
-			return fmt.Errorf("failed to generate bootstrap token: %w", err)
-		}
-
-		if err := talos.CreateBootstrapTokenSecret(ctx, tenantClient.Clientset, bootstrapToken); err != nil {
-			return fmt.Errorf("failed to create bootstrap token in tenant: %w", err)
-		}
-
-		logger.Info("created kubeadm bootstrap token in tenant API server")
-	}
-
-	// Get networking CIDRs
-	podCIDR := "10.244.0.0/16"
-	serviceCIDR := "10.96.0.0/12"
-	if tc.Spec.Networking.PodCIDR != "" {
-		podCIDR = tc.Spec.Networking.PodCIDR
-	}
-	if tc.Spec.Networking.ServiceCIDR != "" {
-		serviceCIDR = tc.Spec.Networking.ServiceCIDR
-	}
-
-	// Get Talos-specific config
-	installDisk := "/dev/vda"
-	var installerImage string
-	if tc.Spec.Workers.MachineTemplate.OS.Talos != nil {
-		if tc.Spec.Workers.MachineTemplate.OS.Talos.InstallDisk != "" {
-			installDisk = tc.Spec.Workers.MachineTemplate.OS.Talos.InstallDisk
-		}
-		installerImage = tc.Spec.Workers.MachineTemplate.OS.Talos.InstallerImage
-	}
-
-	// Generate Talos machine config
-	input := talos.MachineConfigInput{
-		ClusterName:          tc.Name,
-		ControlPlaneEndpoint: controlPlaneEndpoint,
-		ClusterCACert:        clusterCACert,
-		MachineToken:         machineToken,
-		BootstrapToken:       bootstrapToken,
-		OSCACert:             osCACert,
-		PodCIDR:              podCIDR,
-		ServiceCIDR:          serviceCIDR,
-		InstallDisk:          installDisk,
-		InstallerImage:       installerImage,
-	}
-
-	machineConfigData, err := talos.GenerateWorkerConfig(input)
-	if err != nil {
-		return fmt.Errorf("failed to generate Talos machine config: %w", err)
-	}
-
-	// Create the bootstrap Secret in the tenant namespace on the management cluster.
-	// CAPI reads this via dataSecretName on the MachineDeployment.
-	bootstrapSecret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: tenantNS,
-			Labels: map[string]string{
-				butlerv1alpha1.LabelManagedBy: "butler",
-				butlerv1alpha1.LabelTenant:    tc.Name,
-			},
-		},
-		Data: map[string][]byte{
-			"value": machineConfigData,
-		},
-	}
-
-	// Skip owner reference — TC is in butler-tenants namespace but bootstrap Secret
-	// is in the tenant namespace. Cross-namespace owner refs are disallowed.
-	// Labels track ownership instead.
-
-	if err := r.Create(ctx, bootstrapSecret); err != nil {
-		if apierrors.IsAlreadyExists(err) {
-			return nil
-		}
-		return fmt.Errorf("failed to create Talos bootstrap Secret: %w", err)
-	}
-
-	logger.Info("created Talos bootstrap Secret",
-		"name", secretName,
-		"namespace", tenantNS,
-		"endpoint", controlPlaneEndpoint)
-
-	return nil
-}
-
-// validateProviderAccess checks if the TenantCluster's team has access to the ProviderConfig.
-func (r *Reconciler) validateProviderAccess(ctx context.Context, tc *butlerv1alpha1.TenantCluster, pc *butlerv1alpha1.ProviderConfig) error {
-	// Platform-scoped providers are accessible to all teams
-	if pc.Spec.Scope == nil || pc.Spec.Scope.Type == "" || pc.Spec.Scope.Type == butlerv1alpha1.ProviderConfigScopePlatform {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-			metav1.ConditionTrue, butlerv1alpha1.ReasonReady, "platform-scoped provider")
-		return r.validateProviderLimits(ctx, tc, pc)
-	}
-
-	// Team-scoped: verify the TC's team matches the PC's team
-	if pc.Spec.Scope.TeamRef == nil {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-			metav1.ConditionFalse, butlerv1alpha1.ReasonProviderAccessDenied,
-			"provider is team-scoped but has no teamRef")
-		return fmt.Errorf("provider %s is team-scoped but has no teamRef", pc.Name)
-	}
-
-	if tc.Spec.TeamRef == nil {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-			metav1.ConditionFalse, butlerv1alpha1.ReasonProviderAccessDenied,
-			"cluster has no teamRef but provider is team-scoped")
-		return fmt.Errorf("cluster has no teamRef but provider %s is team-scoped to %s",
-			pc.Name, pc.Spec.Scope.TeamRef.Name)
-	}
-
-	if tc.Spec.TeamRef.Name != pc.Spec.Scope.TeamRef.Name {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-			metav1.ConditionFalse, butlerv1alpha1.ReasonProviderAccessDenied,
-			fmt.Sprintf("team %s does not have access to provider %s (scoped to %s)",
-				tc.Spec.TeamRef.Name, pc.Name, pc.Spec.Scope.TeamRef.Name))
-		return fmt.Errorf("team %s does not have access to provider %s (scoped to %s)",
-			tc.Spec.TeamRef.Name, pc.Name, pc.Spec.Scope.TeamRef.Name)
-	}
-
-	r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-		metav1.ConditionTrue, butlerv1alpha1.ReasonReady,
-		fmt.Sprintf("team %s has access to provider %s", tc.Spec.TeamRef.Name, pc.Name))
-
-	// Enforce provider limits (defense-in-depth when webhooks are disabled)
-	return r.validateProviderLimits(ctx, tc, pc)
-}
-
-// validateProviderLimits checks maxClustersPerTeam and maxNodesPerTeam limits from ProviderConfig,
-// then checks team-level resource quotas as defense-in-depth when webhooks are disabled.
-func (r *Reconciler) validateProviderLimits(ctx context.Context, tc *butlerv1alpha1.TenantCluster, pc *butlerv1alpha1.ProviderConfig) error {
-	if pc.Spec.Limits != nil {
-		if pc.Spec.Limits.MaxClustersPerTeam != nil {
-			maxClusters := *pc.Spec.Limits.MaxClustersPerTeam
-			var tcList butlerv1alpha1.TenantClusterList
-			if err := r.List(ctx, &tcList, client.InNamespace(tc.Namespace)); err != nil {
-				return fmt.Errorf("failed to list TenantClusters: %w", err)
-			}
-
-			var existingCount int32
-			for i := range tcList.Items {
-				if tcList.Items[i].Name != tc.Name {
-					existingCount++
-				}
-			}
-
-			if existingCount >= maxClusters {
-				msg := fmt.Sprintf("team namespace %q has %d cluster(s), provider %q limits to %d",
-					tc.Namespace, existingCount, pc.Name, maxClusters)
-				r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-					metav1.ConditionFalse, butlerv1alpha1.ReasonQuotaExceeded, msg)
-				return fmt.Errorf("quota exceeded: %s", msg)
-			}
-		}
-
-		if pc.Spec.Limits.MaxNodesPerTeam != nil {
-			maxNodes := *pc.Spec.Limits.MaxNodesPerTeam
-			var tcList butlerv1alpha1.TenantClusterList
-			if err := r.List(ctx, &tcList, client.InNamespace(tc.Namespace)); err != nil {
-				return fmt.Errorf("failed to list TenantClusters: %w", err)
-			}
-
-			totalNodes := tc.Spec.Workers.Replicas
-			for i := range tcList.Items {
-				if tcList.Items[i].Name != tc.Name {
-					totalNodes += tcList.Items[i].Spec.Workers.Replicas
-				}
-			}
-
-			if totalNodes > maxNodes {
-				msg := fmt.Sprintf("total worker replicas (%d) exceeds provider %q per-team limit (%d)",
-					totalNodes, pc.Name, maxNodes)
-				r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-					metav1.ConditionFalse, butlerv1alpha1.ReasonQuotaExceeded, msg)
-				return fmt.Errorf("quota exceeded: %s", msg)
-			}
-		}
-	}
-
-	// Check team-level resource quotas (defense-in-depth when webhooks are disabled)
-	return r.validateTeamQuotas(ctx, tc)
-}
-
-// validateTeamQuotas checks Team.spec.resourceLimits for maxClusters, maxTotalNodes,
-// and maxNodesPerCluster. CPU/Memory/Storage are checked by the webhook.
-func (r *Reconciler) validateTeamQuotas(ctx context.Context, tc *butlerv1alpha1.TenantCluster) error {
-	if tc.Spec.TeamRef == nil {
-		return nil
-	}
-
-	logger := log.FromContext(ctx)
-
-	team := &butlerv1alpha1.Team{}
-	if err := r.Get(ctx, types.NamespacedName{Name: tc.Spec.TeamRef.Name}, team); err != nil {
-		logger.V(1).Info("could not fetch team for quota check", "team", tc.Spec.TeamRef.Name, "error", err)
-		return nil // Don't block if team can't be fetched
-	}
-
-	limits := team.Spec.ResourceLimits
-	if limits == nil {
-		return nil
-	}
-
-	// Per-cluster node limit
-	if limits.MaxNodesPerCluster != nil {
-		if tc.Spec.Workers.Replicas > *limits.MaxNodesPerCluster {
-			msg := fmt.Sprintf("worker replicas (%d) exceeds team %q per-cluster node limit (%d)",
-				tc.Spec.Workers.Replicas, team.Name, *limits.MaxNodesPerCluster)
-			r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-				metav1.ConditionFalse, butlerv1alpha1.ReasonQuotaExceeded, msg)
-			return fmt.Errorf("quota exceeded: %s", msg)
-		}
-	}
-
-	var tcList butlerv1alpha1.TenantClusterList
-	if err := r.List(ctx, &tcList, client.InNamespace(tc.Namespace)); err != nil {
-		return fmt.Errorf("failed to list TenantClusters: %w", err)
-	}
-
-	// Cluster count limit
-	if limits.MaxClusters != nil {
-		var existingCount int32
-		for i := range tcList.Items {
-			if tcList.Items[i].Name != tc.Name {
-				existingCount++
-			}
-		}
-		if existingCount+1 > *limits.MaxClusters {
-			msg := fmt.Sprintf("team %q has %d cluster(s), team quota limits to %d",
-				team.Name, existingCount, *limits.MaxClusters)
-			r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-				metav1.ConditionFalse, butlerv1alpha1.ReasonQuotaExceeded, msg)
-			return fmt.Errorf("quota exceeded: %s", msg)
-		}
-	}
-
-	// Total nodes limit
-	if limits.MaxTotalNodes != nil {
-		totalNodes := tc.Spec.Workers.Replicas
-		for i := range tcList.Items {
-			if tcList.Items[i].Name != tc.Name {
-				totalNodes += tcList.Items[i].Spec.Workers.Replicas
-			}
-		}
-		if totalNodes > *limits.MaxTotalNodes {
-			msg := fmt.Sprintf("total worker nodes (%d) would exceed team %q node quota (%d)",
-				totalNodes, team.Name, *limits.MaxTotalNodes)
-			r.setCondition(tc, butlerv1alpha1.TenantClusterConditionProviderAccessGranted,
-				metav1.ConditionFalse, butlerv1alpha1.ReasonQuotaExceeded, msg)
-			return fmt.Errorf("quota exceeded: %s", msg)
-		}
-	}
-
-	return nil
-}
-
-// reconcileIPAllocation creates or checks LB IPAllocation for IPAM mode providers.
-// Returns true if IPAM is ready (allocated or not needed), false if waiting.
-func (r *Reconciler) reconcileIPAllocation(ctx context.Context, tc *butlerv1alpha1.TenantCluster, pc *butlerv1alpha1.ProviderConfig) (bool, error) {
-	logger := log.FromContext(ctx)
-
-	// Skip IPAM if not configured
-	if pc.Spec.Network == nil || pc.Spec.Network.Mode != "ipam" {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionNetworkReady,
-			metav1.ConditionTrue, butlerv1alpha1.ReasonReady, "cloud networking mode")
-		return true, nil
-	}
-
-	if len(pc.Spec.Network.PoolRefs) == 0 {
-		r.setCondition(tc, butlerv1alpha1.TenantClusterConditionNetworkReady,
-			metav1.ConditionFalse, butlerv1alpha1.ReasonNetworkNotReady,
-			"IPAM mode but no poolRefs configured on provider")
-		return false, fmt.Errorf("IPAM mode but no poolRefs configured on provider %s", pc.Name)
-	}
-
-	// Check if LB allocation already exists
-	if tc.Status.LBAllocationRef != nil {
-		alloc := &butlerv1alpha1.IPAllocation{}
-		if err := r.Get(ctx, types.NamespacedName{
-			Name:      tc.Status.LBAllocationRef.Name,
-			Namespace: "butler-system",
-		}, alloc); err != nil {
-			if apierrors.IsNotFound(err) {
-				// Allocation was deleted, clear ref and retry
-				tc.Status.LBAllocationRef = nil
-			} else {
-				return false, err
-			}
-		} else {
-			switch alloc.Status.Phase {
-			case butlerv1alpha1.IPAllocationPhaseAllocated:
-				r.setCondition(tc, butlerv1alpha1.TenantClusterConditionNetworkReady,
-					metav1.ConditionTrue, butlerv1alpha1.ReasonReady,
-					fmt.Sprintf("LB IPs allocated: %s", alloc.Status.CIDR))
-				return true, nil
-			case butlerv1alpha1.IPAllocationPhasePending:
-				r.setCondition(tc, butlerv1alpha1.TenantClusterConditionNetworkReady,
-					metav1.ConditionFalse, butlerv1alpha1.ReasonReconciling,
-					"waiting for LB IP allocation")
-				return false, nil
-			case butlerv1alpha1.IPAllocationPhaseFailed:
-				// Delete failed allocation and try next pool
-				logger.Info("LB allocation failed, cleaning up", "allocation", alloc.Name)
-				if err := r.Delete(ctx, alloc); err != nil && !apierrors.IsNotFound(err) {
-					return false, err
-				}
-				tc.Status.LBAllocationRef = nil
-			}
-		}
-	}
-
-	// Create new allocation — try pools in priority order
-	lbCount := r.getInitialLBPoolSize(tc, pc)
-
-	// Enforce quota on initial allocation
-	if pc.Spec.Network.QuotaPerTenant != nil && pc.Spec.Network.QuotaPerTenant.MaxLoadBalancerIPs != nil {
-		maxLB := *pc.Spec.Network.QuotaPerTenant.MaxLoadBalancerIPs
-		if lbCount > maxLB {
-			lbCount = maxLB
-		}
-	}
-
-	allocName := fmt.Sprintf("%s-%s-lb", tc.Namespace, tc.Name)
-
-	for _, poolRef := range pc.Spec.Network.PoolRefs {
-		// Check pool capacity
-		pool := &butlerv1alpha1.NetworkPool{}
-		if err := r.Get(ctx, types.NamespacedName{
-			Name:      poolRef.Name,
-			Namespace: "butler-system",
-		}, pool); err != nil {
-			logger.V(1).Info("pool not found, trying next", "pool", poolRef.Name)
-			continue
-		}
-
-		if pool.Status.AvailableIPs < lbCount {
-			logger.V(1).Info("pool has insufficient capacity, trying next",
-				"pool", poolRef.Name, "available", pool.Status.AvailableIPs, "needed", lbCount)
-			continue
-		}
-
-		// Create IPAllocation
-		alloc := &butlerv1alpha1.IPAllocation{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      allocName,
-				Namespace: "butler-system",
-				Labels: map[string]string{
-					butlerv1alpha1.LabelNetworkPool:    poolRef.Name,
-					butlerv1alpha1.LabelTeam:           tc.Namespace,
-					butlerv1alpha1.LabelTenant:         tc.Name,
-					butlerv1alpha1.LabelAllocationType: "loadbalancer",
-				},
-			},
-			Spec: butlerv1alpha1.IPAllocationSpec{
-				PoolRef:          butlerv1alpha1.LocalObjectReference{Name: poolRef.Name},
-				TenantClusterRef: butlerv1alpha1.NamespacedObjectReference{Name: tc.Name, Namespace: tc.Namespace},
-				Type:             butlerv1alpha1.IPAllocationTypeLoadBalancer,
-				Count:            &lbCount,
-			},
-		}
-
-		if err := r.Create(ctx, alloc); err != nil {
-			if apierrors.IsAlreadyExists(err) {
-				// Re-fetch and check status
-				existing := &butlerv1alpha1.IPAllocation{}
-				if getErr := r.Get(ctx, types.NamespacedName{Name: allocName, Namespace: "butler-system"}, existing); getErr == nil {
-					tc.Status.LBAllocationRef = &butlerv1alpha1.LocalObjectReference{Name: allocName}
-					return false, nil
-				}
-			}
-			return false, fmt.Errorf("failed to create IPAllocation: %w", err)
-		}
-
-		logger.Info("created LB IPAllocation", "allocation", allocName, "pool", poolRef.Name, "count", lbCount)
-		tc.Status.LBAllocationRef = &butlerv1alpha1.LocalObjectReference{Name: allocName}
-		return false, nil // Wait for allocation to be fulfilled
-	}
-
-	// All pools exhausted
-	r.setCondition(tc, butlerv1alpha1.TenantClusterConditionNetworkReady,
-		metav1.ConditionFalse, butlerv1alpha1.ReasonPoolExhausted,
-		"all configured pools are exhausted")
-	return false, nil
-}
-
-func (r *Reconciler) getInitialLBPoolSize(tc *butlerv1alpha1.TenantCluster, pc *butlerv1alpha1.ProviderConfig) int32 {
-	// TC override takes priority
-	if tc.Spec.Networking.LBPoolSize != nil {
-		return *tc.Spec.Networking.LBPoolSize
-	}
-	if pc.Spec.Network != nil && pc.Spec.Network.LoadBalancer != nil {
-		lb := pc.Spec.Network.LoadBalancer
-		if lb.AllocationMode == "elastic" {
-			if lb.InitialPoolSize != nil {
-				return *lb.InitialPoolSize
-			}
-			return 2
-		}
-		if lb.DefaultPoolSize != nil {
-			return *lb.DefaultPoolSize
-		}
-	}
-	return 8
-}
-
-func (r *Reconciler) isElasticIPAM(pc *butlerv1alpha1.ProviderConfig) bool {
-	return pc.Spec.Network != nil &&
-		pc.Spec.Network.Mode == "ipam" &&
-		pc.Spec.Network.LoadBalancer != nil &&
-		pc.Spec.Network.LoadBalancer.AllocationMode == "elastic"
-}
-
-func (r *Reconciler) getGrowthIncrement(pc *butlerv1alpha1.ProviderConfig) int32 {
-	if pc.Spec.Network != nil && pc.Spec.Network.LoadBalancer != nil && pc.Spec.Network.LoadBalancer.GrowthIncrement != nil {
-		return *pc.Spec.Network.LoadBalancer.GrowthIncrement
-	}
-	return 2
-}
-
-// listLBAllocations returns all LB IPAllocations for a given tenant cluster.
-func (r *Reconciler) listLBAllocations(ctx context.Context, tc *butlerv1alpha1.TenantCluster) ([]butlerv1alpha1.IPAllocation, error) {
-	allocList := &butlerv1alpha1.IPAllocationList{}
-	if err := r.List(ctx, allocList, client.InNamespace("butler-system"),
-		client.MatchingLabels{
-			butlerv1alpha1.LabelTeam:           tc.Namespace,
-			butlerv1alpha1.LabelTenant:         tc.Name,
-			butlerv1alpha1.LabelAllocationType: "loadbalancer",
-		}); err != nil {
-		return nil, err
-	}
-	return allocList.Items, nil
-}
-
-// countUsedLBIPs counts the number of LoadBalancer services with assigned IPs on a tenant cluster.
-func (r *Reconciler) countUsedLBIPs(ctx context.Context, tc *tenant.TenantClient) (int32, error) {
-	svcList, err := tc.Clientset.CoreV1().Services("").List(ctx, metav1.ListOptions{})
-	if err != nil {
-		return 0, fmt.Errorf("failed to list services: %w", err)
-	}
-
-	var count int32
-	for _, svc := range svcList.Items {
-		if svc.Spec.Type != corev1.ServiceTypeLoadBalancer {
-			continue
-		}
-		for _, ing := range svc.Status.LoadBalancer.Ingress {
-			if ing.IP != "" {
-				count++
-				break
-			}
-		}
-	}
-	return count, nil
-}
-
-// reconcileElasticIPAM handles dynamic LB IP allocation growth and shrink for Ready clusters.
-func (r *Reconciler) reconcileElasticIPAM(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig) error {
-	if tc.Status.Phase != butlerv1alpha1.TenantClusterPhaseReady {
-		return nil
-	}
-
-	logger := log.FromContext(ctx)
-
-	pc, err := r.getProviderConfig(ctx, tc)
-	if err != nil {
-		return nil // Can't get provider, skip silently
-	}
-
-	if !r.isElasticIPAM(pc) {
-		return nil
-	}
-
-	// Get all LB allocations for this tenant
-	allocs, err := r.listLBAllocations(ctx, tc)
-	if err != nil {
-		return fmt.Errorf("failed to list LB allocations: %w", err)
-	}
-
-	if len(allocs) == 0 {
-		return nil // No allocations yet, initial allocation handles this
-	}
-
-	// Count total allocated and available IPs
-	var totalAllocated int32
-	for _, alloc := range allocs {
-		if alloc.Status.Phase == butlerv1alpha1.IPAllocationPhaseAllocated {
-			if alloc.Spec.Count != nil {
-				totalAllocated += *alloc.Spec.Count
-			}
-		}
-	}
-
-	// Count used IPs from tenant cluster
-	tenantClient, err := r.getTenantClient(ctx, tc)
-	if err != nil {
-		logger.V(1).Info("cannot get tenant client for elastic IPAM, skipping", "error", err)
-		return nil
-	}
-
-	usedIPs, err := r.countUsedLBIPs(ctx, tenantClient)
-	if err != nil {
-		logger.V(1).Info("cannot count used LB IPs, skipping", "error", err)
-		return nil
-	}
-
-	availableIPs := totalAllocated - usedIPs
-	growthIncrement := r.getGrowthIncrement(pc)
-
-	logger.V(1).Info("elastic IPAM status",
-		"totalAllocated", totalAllocated, "usedIPs", usedIPs,
-		"availableIPs", availableIPs, "growthIncrement", growthIncrement)
-
-	// Grow: if fewer than 1 IP is available
-	if availableIPs < 1 {
-		// Check quota
-		if pc.Spec.Network.QuotaPerTenant != nil && pc.Spec.Network.QuotaPerTenant.MaxLoadBalancerIPs != nil {
-			maxLB := *pc.Spec.Network.QuotaPerTenant.MaxLoadBalancerIPs
-			if totalAllocated+growthIncrement > maxLB {
-				logger.Info("elastic IPAM growth blocked by quota",
-					"totalAllocated", totalAllocated, "maxLB", maxLB)
-				return nil
-			}
-		}
-
-		// Find next allocation index
-		nextIdx := len(allocs)
-		allocName := fmt.Sprintf("%s-%s-lb-%d", tc.Namespace, tc.Name, nextIdx)
-
-		// Find a pool with capacity
-		for _, poolRef := range pc.Spec.Network.PoolRefs {
-			pool := &butlerv1alpha1.NetworkPool{}
-			if err := r.Get(ctx, types.NamespacedName{
-				Name: poolRef.Name, Namespace: "butler-system",
-			}, pool); err != nil {
-				continue
-			}
-			if pool.Status.AvailableIPs < growthIncrement {
-				continue
-			}
-
-			alloc := &butlerv1alpha1.IPAllocation{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      allocName,
-					Namespace: "butler-system",
-					Labels: map[string]string{
-						butlerv1alpha1.LabelNetworkPool:    poolRef.Name,
-						butlerv1alpha1.LabelTeam:           tc.Namespace,
-						butlerv1alpha1.LabelTenant:         tc.Name,
-						butlerv1alpha1.LabelAllocationType: "loadbalancer",
-					},
-				},
-				Spec: butlerv1alpha1.IPAllocationSpec{
-					PoolRef:          butlerv1alpha1.LocalObjectReference{Name: poolRef.Name},
-					TenantClusterRef: butlerv1alpha1.NamespacedObjectReference{Name: tc.Name, Namespace: tc.Namespace},
-					Type:             butlerv1alpha1.IPAllocationTypeLoadBalancer,
-					Count:            &growthIncrement,
-				},
-			}
-
-			if err := r.Create(ctx, alloc); err != nil {
-				if apierrors.IsAlreadyExists(err) {
-					return nil // Already created
-				}
-				return fmt.Errorf("failed to create growth allocation: %w", err)
-			}
-			logger.Info("elastic IPAM: created growth allocation",
-				"allocation", allocName, "pool", poolRef.Name, "count", growthIncrement)
-			break
-		}
-	}
-
-	// Shrink: if we have more than 1 allocation and enough spare IPs
-	if len(allocs) > 1 && availableIPs > growthIncrement {
-		// Find the newest allocation that is fully unused and older than 10 minutes
-		var shrinkCandidate *butlerv1alpha1.IPAllocation
-		for i := len(allocs) - 1; i >= 1; i-- { // Skip index 0 (initial allocation)
-			alloc := &allocs[i]
-			if alloc.Status.Phase != butlerv1alpha1.IPAllocationPhaseAllocated {
-				continue
-			}
-			if alloc.Spec.Count == nil {
-				continue
-			}
-			// Only shrink allocations older than 10 minutes to avoid thrashing
-			if time.Since(alloc.CreationTimestamp.Time) < 10*time.Minute {
-				continue
-			}
-			shrinkCandidate = alloc
-			break
-		}
-
-		if shrinkCandidate != nil {
-			logger.Info("elastic IPAM: shrinking unused allocation",
-				"allocation", shrinkCandidate.Name, "availableIPs", availableIPs)
-			if err := r.Delete(ctx, shrinkCandidate); err != nil && !apierrors.IsNotFound(err) {
-				return fmt.Errorf("failed to delete shrink allocation: %w", err)
-			}
-		}
-	}
-
-	// Update MetalLB pool with all current allocated ranges
-	if err := r.updateMetalLBPool(ctx, tc, butlerConfig, allocs); err != nil {
-		logger.Error(err, "failed to update MetalLB pool")
-	}
-
-	return nil
-}
-
-// updateMetalLBPool configures MetalLB on the tenant cluster with all allocated IP ranges.
-func (r *Reconciler) updateMetalLBPool(ctx context.Context, tc *butlerv1alpha1.TenantCluster, butlerConfig *butlerv1alpha1.ButlerConfig, allocs []butlerv1alpha1.IPAllocation) error {
-	var ranges []string
-	for _, alloc := range allocs {
-		if alloc.Status.Phase != butlerv1alpha1.IPAllocationPhaseAllocated {
-			continue
-		}
-		if alloc.Status.StartAddress != "" && alloc.Status.EndAddress != "" {
-			ranges = append(ranges, fmt.Sprintf("%s-%s", alloc.Status.StartAddress, alloc.Status.EndAddress))
-		} else if alloc.Status.CIDR != "" {
-			ranges = append(ranges, alloc.Status.CIDR)
-		}
-	}
-
-	if len(ranges) == 0 {
-		return nil
-	}
-
-	kubeconfig, err := r.getTenantKubeconfig(ctx, tc, butlerConfig)
-	if err != nil {
-		return err
-	}
-
-	return r.Installer.UpdateMetalLBPool(ctx, kubeconfig, ranges)
-}
-
-// cleanupIPAllocations deletes IPAllocations referenced by the TenantCluster
-// and any additional elastic allocations discovered by labels.
-func (r *Reconciler) cleanupIPAllocations(ctx context.Context, tc *butlerv1alpha1.TenantCluster) {
-	logger := log.FromContext(ctx)
-
-	// Delete referenced allocations
-	refs := []*butlerv1alpha1.LocalObjectReference{tc.Status.IPAllocationRef, tc.Status.LBAllocationRef}
-	deleted := make(map[string]bool)
-	for _, ref := range refs {
-		if ref == nil {
-			continue
-		}
-		alloc := &butlerv1alpha1.IPAllocation{}
-		if err := r.Get(ctx, types.NamespacedName{Name: ref.Name, Namespace: "butler-system"}, alloc); err != nil {
-			if !apierrors.IsNotFound(err) {
-				logger.Error(err, "failed to get IPAllocation for cleanup", "name", ref.Name)
-			}
-			continue
-		}
-		if err := r.Delete(ctx, alloc); err != nil && !apierrors.IsNotFound(err) {
-			logger.Error(err, "failed to delete IPAllocation", "name", ref.Name)
-		} else {
-			logger.Info("deleted IPAllocation during cleanup", "name", ref.Name)
-		}
-		deleted[ref.Name] = true
-	}
-
-	// Also clean up any elastic growth allocations discovered by labels
-	allocList := &butlerv1alpha1.IPAllocationList{}
-	if err := r.List(ctx, allocList, client.InNamespace("butler-system"),
-		client.MatchingLabels{
-			butlerv1alpha1.LabelTeam:   tc.Namespace,
-			butlerv1alpha1.LabelTenant: tc.Name,
-		}); err != nil {
-		logger.Error(err, "failed to list IPAllocations for cleanup")
-		return
-	}
-	for i := range allocList.Items {
-		alloc := &allocList.Items[i]
-		if deleted[alloc.Name] {
-			continue
-		}
-		if err := r.Delete(ctx, alloc); err != nil && !apierrors.IsNotFound(err) {
-			logger.Error(err, "failed to delete elastic IPAllocation", "name", alloc.Name)
-		} else {
-			logger.Info("deleted elastic IPAllocation during cleanup", "name", alloc.Name)
-		}
 	}
 }
